Rest: Provide arg description, and refactor some more
