+++ yamcs-api/src/main/java/org/yamcs/protobuf/Commanding.java
@@ -8698,3461 +8698,6 @@ public final class Commanding {
     // @@protoc_insertion_point(class_scope:commanding.ArgumentAssignmentType)
   }
 
-  public interface ValidateCommandRequestOrBuilder
-      extends com.google.protobuf.MessageOrBuilder {
-
-    // repeated .commanding.CommandType command = 1;
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    java.util.List<org.yamcs.protobuf.Commanding.CommandType> 
-        getCommandList();
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.CommandType getCommand(int index);
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    int getCommandCount();
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    java.util.List<? extends org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-        getCommandOrBuilderList();
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.CommandTypeOrBuilder getCommandOrBuilder(
-        int index);
-  }
-  /**
-   * Protobuf type {@code commanding.ValidateCommandRequest}
-   */
-  public static final class ValidateCommandRequest extends
-      com.google.protobuf.GeneratedMessage
-      implements ValidateCommandRequestOrBuilder {
-    // Use ValidateCommandRequest.newBuilder() to construct.
-    private ValidateCommandRequest(com.google.protobuf.GeneratedMessage.Builder<?> builder) {
-      super(builder);
-      this.unknownFields = builder.getUnknownFields();
-    }
-    private ValidateCommandRequest(boolean noInit) { this.unknownFields = com.google.protobuf.UnknownFieldSet.getDefaultInstance(); }
-
-    private static final ValidateCommandRequest defaultInstance;
-    public static ValidateCommandRequest getDefaultInstance() {
-      return defaultInstance;
-    }
-
-    public ValidateCommandRequest getDefaultInstanceForType() {
-      return defaultInstance;
-    }
-
-    private final com.google.protobuf.UnknownFieldSet unknownFields;
-    @java.lang.Override
-    public final com.google.protobuf.UnknownFieldSet
-        getUnknownFields() {
-      return this.unknownFields;
-    }
-    private ValidateCommandRequest(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      initFields();
-      int mutable_bitField0_ = 0;
-      com.google.protobuf.UnknownFieldSet.Builder unknownFields =
-          com.google.protobuf.UnknownFieldSet.newBuilder();
-      try {
-        boolean done = false;
-        while (!done) {
-          int tag = input.readTag();
-          switch (tag) {
-            case 0:
-              done = true;
-              break;
-            default: {
-              if (!parseUnknownField(input, unknownFields,
-                                     extensionRegistry, tag)) {
-                done = true;
-              }
-              break;
-            }
-            case 10: {
-              if (!((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-                command_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.CommandType>();
-                mutable_bitField0_ |= 0x00000001;
-              }
-              command_.add(input.readMessage(org.yamcs.protobuf.Commanding.CommandType.PARSER, extensionRegistry));
-              break;
-            }
-          }
-        }
-      } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-        throw e.setUnfinishedMessage(this);
-      } catch (java.io.IOException e) {
-        throw new com.google.protobuf.InvalidProtocolBufferException(
-            e.getMessage()).setUnfinishedMessage(this);
-      } finally {
-        if (((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-          command_ = java.util.Collections.unmodifiableList(command_);
-        }
-        this.unknownFields = unknownFields.build();
-        makeExtensionsImmutable();
-      }
-    }
-    public static final com.google.protobuf.Descriptors.Descriptor
-        getDescriptor() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandRequest_descriptor;
-    }
-
-    protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-        internalGetFieldAccessorTable() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandRequest_fieldAccessorTable
-          .ensureFieldAccessorsInitialized(
-              org.yamcs.protobuf.Commanding.ValidateCommandRequest.class, org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder.class);
-    }
-
-    public static com.google.protobuf.Parser<ValidateCommandRequest> PARSER =
-        new com.google.protobuf.AbstractParser<ValidateCommandRequest>() {
-      public ValidateCommandRequest parsePartialFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return new ValidateCommandRequest(input, extensionRegistry);
-      }
-    };
-
-    @java.lang.Override
-    public com.google.protobuf.Parser<ValidateCommandRequest> getParserForType() {
-      return PARSER;
-    }
-
-    // repeated .commanding.CommandType command = 1;
-    public static final int COMMAND_FIELD_NUMBER = 1;
-    private java.util.List<org.yamcs.protobuf.Commanding.CommandType> command_;
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public java.util.List<org.yamcs.protobuf.Commanding.CommandType> getCommandList() {
-      return command_;
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public java.util.List<? extends org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-        getCommandOrBuilderList() {
-      return command_;
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public int getCommandCount() {
-      return command_.size();
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.CommandType getCommand(int index) {
-      return command_.get(index);
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.CommandTypeOrBuilder getCommandOrBuilder(
-        int index) {
-      return command_.get(index);
-    }
-
-    private void initFields() {
-      command_ = java.util.Collections.emptyList();
-    }
-    private byte memoizedIsInitialized = -1;
-    public final boolean isInitialized() {
-      byte isInitialized = memoizedIsInitialized;
-      if (isInitialized != -1) return isInitialized == 1;
-
-      for (int i = 0; i < getCommandCount(); i++) {
-        if (!getCommand(i).isInitialized()) {
-          memoizedIsInitialized = 0;
-          return false;
-        }
-      }
-      memoizedIsInitialized = 1;
-      return true;
-    }
-
-    public void writeTo(com.google.protobuf.CodedOutputStream output)
-                        throws java.io.IOException {
-      getSerializedSize();
-      for (int i = 0; i < command_.size(); i++) {
-        output.writeMessage(1, command_.get(i));
-      }
-      getUnknownFields().writeTo(output);
-    }
-
-    private int memoizedSerializedSize = -1;
-    public int getSerializedSize() {
-      int size = memoizedSerializedSize;
-      if (size != -1) return size;
-
-      size = 0;
-      for (int i = 0; i < command_.size(); i++) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeMessageSize(1, command_.get(i));
-      }
-      size += getUnknownFields().getSerializedSize();
-      memoizedSerializedSize = size;
-      return size;
-    }
-
-    private static final long serialVersionUID = 0L;
-    @java.lang.Override
-    protected java.lang.Object writeReplace()
-        throws java.io.ObjectStreamException {
-      return super.writeReplace();
-    }
-
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(
-        com.google.protobuf.ByteString data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(
-        com.google.protobuf.ByteString data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(byte[] data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(
-        byte[] data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseDelimitedFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseDelimitedFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(
-        com.google.protobuf.CodedInputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandRequest parseFrom(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-
-    public static Builder newBuilder() { return Builder.create(); }
-    public Builder newBuilderForType() { return newBuilder(); }
-    public static Builder newBuilder(org.yamcs.protobuf.Commanding.ValidateCommandRequest prototype) {
-      return newBuilder().mergeFrom(prototype);
-    }
-    public Builder toBuilder() { return newBuilder(this); }
-
-    @java.lang.Override
-    protected Builder newBuilderForType(
-        com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-      Builder builder = new Builder(parent);
-      return builder;
-    }
-    /**
-     * Protobuf type {@code commanding.ValidateCommandRequest}
-     */
-    public static final class Builder extends
-        com.google.protobuf.GeneratedMessage.Builder<Builder>
-       implements org.yamcs.protobuf.Commanding.ValidateCommandRequestOrBuilder {
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandRequest_descriptor;
-      }
-
-      protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandRequest_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                org.yamcs.protobuf.Commanding.ValidateCommandRequest.class, org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder.class);
-      }
-
-      // Construct using org.yamcs.protobuf.Commanding.ValidateCommandRequest.newBuilder()
-      private Builder() {
-        maybeForceBuilderInitialization();
-      }
-
-      private Builder(
-          com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-        super(parent);
-        maybeForceBuilderInitialization();
-      }
-      private void maybeForceBuilderInitialization() {
-        if (com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders) {
-          getCommandFieldBuilder();
-        }
-      }
-      private static Builder create() {
-        return new Builder();
-      }
-
-      public Builder clear() {
-        super.clear();
-        if (commandBuilder_ == null) {
-          command_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-        } else {
-          commandBuilder_.clear();
-        }
-        return this;
-      }
-
-      public Builder clone() {
-        return create().mergeFrom(buildPartial());
-      }
-
-      public com.google.protobuf.Descriptors.Descriptor
-          getDescriptorForType() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandRequest_descriptor;
-      }
-
-      public org.yamcs.protobuf.Commanding.ValidateCommandRequest getDefaultInstanceForType() {
-        return org.yamcs.protobuf.Commanding.ValidateCommandRequest.getDefaultInstance();
-      }
-
-      public org.yamcs.protobuf.Commanding.ValidateCommandRequest build() {
-        org.yamcs.protobuf.Commanding.ValidateCommandRequest result = buildPartial();
-        if (!result.isInitialized()) {
-          throw newUninitializedMessageException(result);
-        }
-        return result;
-      }
-
-      public org.yamcs.protobuf.Commanding.ValidateCommandRequest buildPartial() {
-        org.yamcs.protobuf.Commanding.ValidateCommandRequest result = new org.yamcs.protobuf.Commanding.ValidateCommandRequest(this);
-        int from_bitField0_ = bitField0_;
-        if (commandBuilder_ == null) {
-          if (((bitField0_ & 0x00000001) == 0x00000001)) {
-            command_ = java.util.Collections.unmodifiableList(command_);
-            bitField0_ = (bitField0_ & ~0x00000001);
-          }
-          result.command_ = command_;
-        } else {
-          result.command_ = commandBuilder_.build();
-        }
-        onBuilt();
-        return result;
-      }
-
-      public Builder mergeFrom(com.google.protobuf.Message other) {
-        if (other instanceof org.yamcs.protobuf.Commanding.ValidateCommandRequest) {
-          return mergeFrom((org.yamcs.protobuf.Commanding.ValidateCommandRequest)other);
-        } else {
-          super.mergeFrom(other);
-          return this;
-        }
-      }
-
-      public Builder mergeFrom(org.yamcs.protobuf.Commanding.ValidateCommandRequest other) {
-        if (other == org.yamcs.protobuf.Commanding.ValidateCommandRequest.getDefaultInstance()) return this;
-        if (commandBuilder_ == null) {
-          if (!other.command_.isEmpty()) {
-            if (command_.isEmpty()) {
-              command_ = other.command_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-            } else {
-              ensureCommandIsMutable();
-              command_.addAll(other.command_);
-            }
-            onChanged();
-          }
-        } else {
-          if (!other.command_.isEmpty()) {
-            if (commandBuilder_.isEmpty()) {
-              commandBuilder_.dispose();
-              commandBuilder_ = null;
-              command_ = other.command_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-              commandBuilder_ = 
-                com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders ?
-                   getCommandFieldBuilder() : null;
-            } else {
-              commandBuilder_.addAllMessages(other.command_);
-            }
-          }
-        }
-        this.mergeUnknownFields(other.getUnknownFields());
-        return this;
-      }
-
-      public final boolean isInitialized() {
-        for (int i = 0; i < getCommandCount(); i++) {
-          if (!getCommand(i).isInitialized()) {
-            
-            return false;
-          }
-        }
-        return true;
-      }
-
-      public Builder mergeFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        org.yamcs.protobuf.Commanding.ValidateCommandRequest parsedMessage = null;
-        try {
-          parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          parsedMessage = (org.yamcs.protobuf.Commanding.ValidateCommandRequest) e.getUnfinishedMessage();
-          throw e;
-        } finally {
-          if (parsedMessage != null) {
-            mergeFrom(parsedMessage);
-          }
-        }
-        return this;
-      }
-      private int bitField0_;
-
-      // repeated .commanding.CommandType command = 1;
-      private java.util.List<org.yamcs.protobuf.Commanding.CommandType> command_ =
-        java.util.Collections.emptyList();
-      private void ensureCommandIsMutable() {
-        if (!((bitField0_ & 0x00000001) == 0x00000001)) {
-          command_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.CommandType>(command_);
-          bitField0_ |= 0x00000001;
-         }
-      }
-
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.CommandType, org.yamcs.protobuf.Commanding.CommandType.Builder, org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> commandBuilder_;
-
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.CommandType> getCommandList() {
-        if (commandBuilder_ == null) {
-          return java.util.Collections.unmodifiableList(command_);
-        } else {
-          return commandBuilder_.getMessageList();
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public int getCommandCount() {
-        if (commandBuilder_ == null) {
-          return command_.size();
-        } else {
-          return commandBuilder_.getCount();
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType getCommand(int index) {
-        if (commandBuilder_ == null) {
-          return command_.get(index);
-        } else {
-          return commandBuilder_.getMessage(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder setCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType value) {
-        if (commandBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandIsMutable();
-          command_.set(index, value);
-          onChanged();
-        } else {
-          commandBuilder_.setMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder setCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType.Builder builderForValue) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.set(index, builderForValue.build());
-          onChanged();
-        } else {
-          commandBuilder_.setMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(org.yamcs.protobuf.Commanding.CommandType value) {
-        if (commandBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandIsMutable();
-          command_.add(value);
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType value) {
-        if (commandBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandIsMutable();
-          command_.add(index, value);
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(
-          org.yamcs.protobuf.Commanding.CommandType.Builder builderForValue) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.add(builderForValue.build());
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType.Builder builderForValue) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.add(index, builderForValue.build());
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addAllCommand(
-          java.lang.Iterable<? extends org.yamcs.protobuf.Commanding.CommandType> values) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          super.addAll(values, command_);
-          onChanged();
-        } else {
-          commandBuilder_.addAllMessages(values);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder clearCommand() {
-        if (commandBuilder_ == null) {
-          command_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-          onChanged();
-        } else {
-          commandBuilder_.clear();
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder removeCommand(int index) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.remove(index);
-          onChanged();
-        } else {
-          commandBuilder_.remove(index);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType.Builder getCommandBuilder(
-          int index) {
-        return getCommandFieldBuilder().getBuilder(index);
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandTypeOrBuilder getCommandOrBuilder(
-          int index) {
-        if (commandBuilder_ == null) {
-          return command_.get(index);  } else {
-          return commandBuilder_.getMessageOrBuilder(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public java.util.List<? extends org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-           getCommandOrBuilderList() {
-        if (commandBuilder_ != null) {
-          return commandBuilder_.getMessageOrBuilderList();
-        } else {
-          return java.util.Collections.unmodifiableList(command_);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType.Builder addCommandBuilder() {
-        return getCommandFieldBuilder().addBuilder(
-            org.yamcs.protobuf.Commanding.CommandType.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType.Builder addCommandBuilder(
-          int index) {
-        return getCommandFieldBuilder().addBuilder(
-            index, org.yamcs.protobuf.Commanding.CommandType.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.CommandType.Builder> 
-           getCommandBuilderList() {
-        return getCommandFieldBuilder().getBuilderList();
-      }
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.CommandType, org.yamcs.protobuf.Commanding.CommandType.Builder, org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-          getCommandFieldBuilder() {
-        if (commandBuilder_ == null) {
-          commandBuilder_ = new com.google.protobuf.RepeatedFieldBuilder<
-              org.yamcs.protobuf.Commanding.CommandType, org.yamcs.protobuf.Commanding.CommandType.Builder, org.yamcs.protobuf.Commanding.CommandTypeOrBuilder>(
-                  command_,
-                  ((bitField0_ & 0x00000001) == 0x00000001),
-                  getParentForChildren(),
-                  isClean());
-          command_ = null;
-        }
-        return commandBuilder_;
-      }
-
-      // @@protoc_insertion_point(builder_scope:commanding.ValidateCommandRequest)
-    }
-
-    static {
-      defaultInstance = new ValidateCommandRequest(true);
-      defaultInstance.initFields();
-    }
-
-    // @@protoc_insertion_point(class_scope:commanding.ValidateCommandRequest)
-  }
-
-  public interface ValidateCommandResponseOrBuilder
-      extends com.google.protobuf.MessageOrBuilder {
-
-    // repeated .commanding.CommandSignificance commandSignificance = 1;
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    java.util.List<org.yamcs.protobuf.Commanding.CommandSignificance> 
-        getCommandSignificanceList();
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.CommandSignificance getCommandSignificance(int index);
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    int getCommandSignificanceCount();
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    java.util.List<? extends org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder> 
-        getCommandSignificanceOrBuilderList();
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder getCommandSignificanceOrBuilder(
-        int index);
-  }
-  /**
-   * Protobuf type {@code commanding.ValidateCommandResponse}
-   */
-  public static final class ValidateCommandResponse extends
-      com.google.protobuf.GeneratedMessage
-      implements ValidateCommandResponseOrBuilder {
-    // Use ValidateCommandResponse.newBuilder() to construct.
-    private ValidateCommandResponse(com.google.protobuf.GeneratedMessage.Builder<?> builder) {
-      super(builder);
-      this.unknownFields = builder.getUnknownFields();
-    }
-    private ValidateCommandResponse(boolean noInit) { this.unknownFields = com.google.protobuf.UnknownFieldSet.getDefaultInstance(); }
-
-    private static final ValidateCommandResponse defaultInstance;
-    public static ValidateCommandResponse getDefaultInstance() {
-      return defaultInstance;
-    }
-
-    public ValidateCommandResponse getDefaultInstanceForType() {
-      return defaultInstance;
-    }
-
-    private final com.google.protobuf.UnknownFieldSet unknownFields;
-    @java.lang.Override
-    public final com.google.protobuf.UnknownFieldSet
-        getUnknownFields() {
-      return this.unknownFields;
-    }
-    private ValidateCommandResponse(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      initFields();
-      int mutable_bitField0_ = 0;
-      com.google.protobuf.UnknownFieldSet.Builder unknownFields =
-          com.google.protobuf.UnknownFieldSet.newBuilder();
-      try {
-        boolean done = false;
-        while (!done) {
-          int tag = input.readTag();
-          switch (tag) {
-            case 0:
-              done = true;
-              break;
-            default: {
-              if (!parseUnknownField(input, unknownFields,
-                                     extensionRegistry, tag)) {
-                done = true;
-              }
-              break;
-            }
-            case 10: {
-              if (!((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-                commandSignificance_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.CommandSignificance>();
-                mutable_bitField0_ |= 0x00000001;
-              }
-              commandSignificance_.add(input.readMessage(org.yamcs.protobuf.Commanding.CommandSignificance.PARSER, extensionRegistry));
-              break;
-            }
-          }
-        }
-      } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-        throw e.setUnfinishedMessage(this);
-      } catch (java.io.IOException e) {
-        throw new com.google.protobuf.InvalidProtocolBufferException(
-            e.getMessage()).setUnfinishedMessage(this);
-      } finally {
-        if (((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-          commandSignificance_ = java.util.Collections.unmodifiableList(commandSignificance_);
-        }
-        this.unknownFields = unknownFields.build();
-        makeExtensionsImmutable();
-      }
-    }
-    public static final com.google.protobuf.Descriptors.Descriptor
-        getDescriptor() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandResponse_descriptor;
-    }
-
-    protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-        internalGetFieldAccessorTable() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandResponse_fieldAccessorTable
-          .ensureFieldAccessorsInitialized(
-              org.yamcs.protobuf.Commanding.ValidateCommandResponse.class, org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder.class);
-    }
-
-    public static com.google.protobuf.Parser<ValidateCommandResponse> PARSER =
-        new com.google.protobuf.AbstractParser<ValidateCommandResponse>() {
-      public ValidateCommandResponse parsePartialFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return new ValidateCommandResponse(input, extensionRegistry);
-      }
-    };
-
-    @java.lang.Override
-    public com.google.protobuf.Parser<ValidateCommandResponse> getParserForType() {
-      return PARSER;
-    }
-
-    // repeated .commanding.CommandSignificance commandSignificance = 1;
-    public static final int COMMANDSIGNIFICANCE_FIELD_NUMBER = 1;
-    private java.util.List<org.yamcs.protobuf.Commanding.CommandSignificance> commandSignificance_;
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    public java.util.List<org.yamcs.protobuf.Commanding.CommandSignificance> getCommandSignificanceList() {
-      return commandSignificance_;
-    }
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    public java.util.List<? extends org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder> 
-        getCommandSignificanceOrBuilderList() {
-      return commandSignificance_;
-    }
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    public int getCommandSignificanceCount() {
-      return commandSignificance_.size();
-    }
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.CommandSignificance getCommandSignificance(int index) {
-      return commandSignificance_.get(index);
-    }
-    /**
-     * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder getCommandSignificanceOrBuilder(
-        int index) {
-      return commandSignificance_.get(index);
-    }
-
-    private void initFields() {
-      commandSignificance_ = java.util.Collections.emptyList();
-    }
-    private byte memoizedIsInitialized = -1;
-    public final boolean isInitialized() {
-      byte isInitialized = memoizedIsInitialized;
-      if (isInitialized != -1) return isInitialized == 1;
-
-      memoizedIsInitialized = 1;
-      return true;
-    }
-
-    public void writeTo(com.google.protobuf.CodedOutputStream output)
-                        throws java.io.IOException {
-      getSerializedSize();
-      for (int i = 0; i < commandSignificance_.size(); i++) {
-        output.writeMessage(1, commandSignificance_.get(i));
-      }
-      getUnknownFields().writeTo(output);
-    }
-
-    private int memoizedSerializedSize = -1;
-    public int getSerializedSize() {
-      int size = memoizedSerializedSize;
-      if (size != -1) return size;
-
-      size = 0;
-      for (int i = 0; i < commandSignificance_.size(); i++) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeMessageSize(1, commandSignificance_.get(i));
-      }
-      size += getUnknownFields().getSerializedSize();
-      memoizedSerializedSize = size;
-      return size;
-    }
-
-    private static final long serialVersionUID = 0L;
-    @java.lang.Override
-    protected java.lang.Object writeReplace()
-        throws java.io.ObjectStreamException {
-      return super.writeReplace();
-    }
-
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(
-        com.google.protobuf.ByteString data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(
-        com.google.protobuf.ByteString data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(byte[] data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(
-        byte[] data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseDelimitedFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseDelimitedFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(
-        com.google.protobuf.CodedInputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.ValidateCommandResponse parseFrom(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-
-    public static Builder newBuilder() { return Builder.create(); }
-    public Builder newBuilderForType() { return newBuilder(); }
-    public static Builder newBuilder(org.yamcs.protobuf.Commanding.ValidateCommandResponse prototype) {
-      return newBuilder().mergeFrom(prototype);
-    }
-    public Builder toBuilder() { return newBuilder(this); }
-
-    @java.lang.Override
-    protected Builder newBuilderForType(
-        com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-      Builder builder = new Builder(parent);
-      return builder;
-    }
-    /**
-     * Protobuf type {@code commanding.ValidateCommandResponse}
-     */
-    public static final class Builder extends
-        com.google.protobuf.GeneratedMessage.Builder<Builder>
-       implements org.yamcs.protobuf.Commanding.ValidateCommandResponseOrBuilder {
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandResponse_descriptor;
-      }
-
-      protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandResponse_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                org.yamcs.protobuf.Commanding.ValidateCommandResponse.class, org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder.class);
-      }
-
-      // Construct using org.yamcs.protobuf.Commanding.ValidateCommandResponse.newBuilder()
-      private Builder() {
-        maybeForceBuilderInitialization();
-      }
-
-      private Builder(
-          com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-        super(parent);
-        maybeForceBuilderInitialization();
-      }
-      private void maybeForceBuilderInitialization() {
-        if (com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders) {
-          getCommandSignificanceFieldBuilder();
-        }
-      }
-      private static Builder create() {
-        return new Builder();
-      }
-
-      public Builder clear() {
-        super.clear();
-        if (commandSignificanceBuilder_ == null) {
-          commandSignificance_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-        } else {
-          commandSignificanceBuilder_.clear();
-        }
-        return this;
-      }
-
-      public Builder clone() {
-        return create().mergeFrom(buildPartial());
-      }
-
-      public com.google.protobuf.Descriptors.Descriptor
-          getDescriptorForType() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_ValidateCommandResponse_descriptor;
-      }
-
-      public org.yamcs.protobuf.Commanding.ValidateCommandResponse getDefaultInstanceForType() {
-        return org.yamcs.protobuf.Commanding.ValidateCommandResponse.getDefaultInstance();
-      }
-
-      public org.yamcs.protobuf.Commanding.ValidateCommandResponse build() {
-        org.yamcs.protobuf.Commanding.ValidateCommandResponse result = buildPartial();
-        if (!result.isInitialized()) {
-          throw newUninitializedMessageException(result);
-        }
-        return result;
-      }
-
-      public org.yamcs.protobuf.Commanding.ValidateCommandResponse buildPartial() {
-        org.yamcs.protobuf.Commanding.ValidateCommandResponse result = new org.yamcs.protobuf.Commanding.ValidateCommandResponse(this);
-        int from_bitField0_ = bitField0_;
-        if (commandSignificanceBuilder_ == null) {
-          if (((bitField0_ & 0x00000001) == 0x00000001)) {
-            commandSignificance_ = java.util.Collections.unmodifiableList(commandSignificance_);
-            bitField0_ = (bitField0_ & ~0x00000001);
-          }
-          result.commandSignificance_ = commandSignificance_;
-        } else {
-          result.commandSignificance_ = commandSignificanceBuilder_.build();
-        }
-        onBuilt();
-        return result;
-      }
-
-      public Builder mergeFrom(com.google.protobuf.Message other) {
-        if (other instanceof org.yamcs.protobuf.Commanding.ValidateCommandResponse) {
-          return mergeFrom((org.yamcs.protobuf.Commanding.ValidateCommandResponse)other);
-        } else {
-          super.mergeFrom(other);
-          return this;
-        }
-      }
-
-      public Builder mergeFrom(org.yamcs.protobuf.Commanding.ValidateCommandResponse other) {
-        if (other == org.yamcs.protobuf.Commanding.ValidateCommandResponse.getDefaultInstance()) return this;
-        if (commandSignificanceBuilder_ == null) {
-          if (!other.commandSignificance_.isEmpty()) {
-            if (commandSignificance_.isEmpty()) {
-              commandSignificance_ = other.commandSignificance_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-            } else {
-              ensureCommandSignificanceIsMutable();
-              commandSignificance_.addAll(other.commandSignificance_);
-            }
-            onChanged();
-          }
-        } else {
-          if (!other.commandSignificance_.isEmpty()) {
-            if (commandSignificanceBuilder_.isEmpty()) {
-              commandSignificanceBuilder_.dispose();
-              commandSignificanceBuilder_ = null;
-              commandSignificance_ = other.commandSignificance_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-              commandSignificanceBuilder_ = 
-                com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders ?
-                   getCommandSignificanceFieldBuilder() : null;
-            } else {
-              commandSignificanceBuilder_.addAllMessages(other.commandSignificance_);
-            }
-          }
-        }
-        this.mergeUnknownFields(other.getUnknownFields());
-        return this;
-      }
-
-      public final boolean isInitialized() {
-        return true;
-      }
-
-      public Builder mergeFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        org.yamcs.protobuf.Commanding.ValidateCommandResponse parsedMessage = null;
-        try {
-          parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          parsedMessage = (org.yamcs.protobuf.Commanding.ValidateCommandResponse) e.getUnfinishedMessage();
-          throw e;
-        } finally {
-          if (parsedMessage != null) {
-            mergeFrom(parsedMessage);
-          }
-        }
-        return this;
-      }
-      private int bitField0_;
-
-      // repeated .commanding.CommandSignificance commandSignificance = 1;
-      private java.util.List<org.yamcs.protobuf.Commanding.CommandSignificance> commandSignificance_ =
-        java.util.Collections.emptyList();
-      private void ensureCommandSignificanceIsMutable() {
-        if (!((bitField0_ & 0x00000001) == 0x00000001)) {
-          commandSignificance_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.CommandSignificance>(commandSignificance_);
-          bitField0_ |= 0x00000001;
-         }
-      }
-
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.CommandSignificance, org.yamcs.protobuf.Commanding.CommandSignificance.Builder, org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder> commandSignificanceBuilder_;
-
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.CommandSignificance> getCommandSignificanceList() {
-        if (commandSignificanceBuilder_ == null) {
-          return java.util.Collections.unmodifiableList(commandSignificance_);
-        } else {
-          return commandSignificanceBuilder_.getMessageList();
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public int getCommandSignificanceCount() {
-        if (commandSignificanceBuilder_ == null) {
-          return commandSignificance_.size();
-        } else {
-          return commandSignificanceBuilder_.getCount();
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandSignificance getCommandSignificance(int index) {
-        if (commandSignificanceBuilder_ == null) {
-          return commandSignificance_.get(index);
-        } else {
-          return commandSignificanceBuilder_.getMessage(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder setCommandSignificance(
-          int index, org.yamcs.protobuf.Commanding.CommandSignificance value) {
-        if (commandSignificanceBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.set(index, value);
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.setMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder setCommandSignificance(
-          int index, org.yamcs.protobuf.Commanding.CommandSignificance.Builder builderForValue) {
-        if (commandSignificanceBuilder_ == null) {
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.set(index, builderForValue.build());
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.setMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder addCommandSignificance(org.yamcs.protobuf.Commanding.CommandSignificance value) {
-        if (commandSignificanceBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.add(value);
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.addMessage(value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder addCommandSignificance(
-          int index, org.yamcs.protobuf.Commanding.CommandSignificance value) {
-        if (commandSignificanceBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.add(index, value);
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.addMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder addCommandSignificance(
-          org.yamcs.protobuf.Commanding.CommandSignificance.Builder builderForValue) {
-        if (commandSignificanceBuilder_ == null) {
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.add(builderForValue.build());
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.addMessage(builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder addCommandSignificance(
-          int index, org.yamcs.protobuf.Commanding.CommandSignificance.Builder builderForValue) {
-        if (commandSignificanceBuilder_ == null) {
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.add(index, builderForValue.build());
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.addMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder addAllCommandSignificance(
-          java.lang.Iterable<? extends org.yamcs.protobuf.Commanding.CommandSignificance> values) {
-        if (commandSignificanceBuilder_ == null) {
-          ensureCommandSignificanceIsMutable();
-          super.addAll(values, commandSignificance_);
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.addAllMessages(values);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder clearCommandSignificance() {
-        if (commandSignificanceBuilder_ == null) {
-          commandSignificance_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.clear();
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public Builder removeCommandSignificance(int index) {
-        if (commandSignificanceBuilder_ == null) {
-          ensureCommandSignificanceIsMutable();
-          commandSignificance_.remove(index);
-          onChanged();
-        } else {
-          commandSignificanceBuilder_.remove(index);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandSignificance.Builder getCommandSignificanceBuilder(
-          int index) {
-        return getCommandSignificanceFieldBuilder().getBuilder(index);
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder getCommandSignificanceOrBuilder(
-          int index) {
-        if (commandSignificanceBuilder_ == null) {
-          return commandSignificance_.get(index);  } else {
-          return commandSignificanceBuilder_.getMessageOrBuilder(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public java.util.List<? extends org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder> 
-           getCommandSignificanceOrBuilderList() {
-        if (commandSignificanceBuilder_ != null) {
-          return commandSignificanceBuilder_.getMessageOrBuilderList();
-        } else {
-          return java.util.Collections.unmodifiableList(commandSignificance_);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandSignificance.Builder addCommandSignificanceBuilder() {
-        return getCommandSignificanceFieldBuilder().addBuilder(
-            org.yamcs.protobuf.Commanding.CommandSignificance.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandSignificance.Builder addCommandSignificanceBuilder(
-          int index) {
-        return getCommandSignificanceFieldBuilder().addBuilder(
-            index, org.yamcs.protobuf.Commanding.CommandSignificance.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.CommandSignificance commandSignificance = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.CommandSignificance.Builder> 
-           getCommandSignificanceBuilderList() {
-        return getCommandSignificanceFieldBuilder().getBuilderList();
-      }
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.CommandSignificance, org.yamcs.protobuf.Commanding.CommandSignificance.Builder, org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder> 
-          getCommandSignificanceFieldBuilder() {
-        if (commandSignificanceBuilder_ == null) {
-          commandSignificanceBuilder_ = new com.google.protobuf.RepeatedFieldBuilder<
-              org.yamcs.protobuf.Commanding.CommandSignificance, org.yamcs.protobuf.Commanding.CommandSignificance.Builder, org.yamcs.protobuf.Commanding.CommandSignificanceOrBuilder>(
-                  commandSignificance_,
-                  ((bitField0_ & 0x00000001) == 0x00000001),
-                  getParentForChildren(),
-                  isClean());
-          commandSignificance_ = null;
-        }
-        return commandSignificanceBuilder_;
-      }
-
-      // @@protoc_insertion_point(builder_scope:commanding.ValidateCommandResponse)
-    }
-
-    static {
-      defaultInstance = new ValidateCommandResponse(true);
-      defaultInstance.initFields();
-    }
-
-    // @@protoc_insertion_point(class_scope:commanding.ValidateCommandResponse)
-  }
-
-  public interface SendCommandRequestOrBuilder
-      extends com.google.protobuf.MessageOrBuilder {
-
-    // repeated .commanding.CommandType command = 1;
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    java.util.List<org.yamcs.protobuf.Commanding.CommandType> 
-        getCommandList();
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.CommandType getCommand(int index);
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    int getCommandCount();
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    java.util.List<? extends org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-        getCommandOrBuilderList();
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.CommandTypeOrBuilder getCommandOrBuilder(
-        int index);
-  }
-  /**
-   * Protobuf type {@code commanding.SendCommandRequest}
-   */
-  public static final class SendCommandRequest extends
-      com.google.protobuf.GeneratedMessage
-      implements SendCommandRequestOrBuilder {
-    // Use SendCommandRequest.newBuilder() to construct.
-    private SendCommandRequest(com.google.protobuf.GeneratedMessage.Builder<?> builder) {
-      super(builder);
-      this.unknownFields = builder.getUnknownFields();
-    }
-    private SendCommandRequest(boolean noInit) { this.unknownFields = com.google.protobuf.UnknownFieldSet.getDefaultInstance(); }
-
-    private static final SendCommandRequest defaultInstance;
-    public static SendCommandRequest getDefaultInstance() {
-      return defaultInstance;
-    }
-
-    public SendCommandRequest getDefaultInstanceForType() {
-      return defaultInstance;
-    }
-
-    private final com.google.protobuf.UnknownFieldSet unknownFields;
-    @java.lang.Override
-    public final com.google.protobuf.UnknownFieldSet
-        getUnknownFields() {
-      return this.unknownFields;
-    }
-    private SendCommandRequest(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      initFields();
-      int mutable_bitField0_ = 0;
-      com.google.protobuf.UnknownFieldSet.Builder unknownFields =
-          com.google.protobuf.UnknownFieldSet.newBuilder();
-      try {
-        boolean done = false;
-        while (!done) {
-          int tag = input.readTag();
-          switch (tag) {
-            case 0:
-              done = true;
-              break;
-            default: {
-              if (!parseUnknownField(input, unknownFields,
-                                     extensionRegistry, tag)) {
-                done = true;
-              }
-              break;
-            }
-            case 10: {
-              if (!((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-                command_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.CommandType>();
-                mutable_bitField0_ |= 0x00000001;
-              }
-              command_.add(input.readMessage(org.yamcs.protobuf.Commanding.CommandType.PARSER, extensionRegistry));
-              break;
-            }
-          }
-        }
-      } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-        throw e.setUnfinishedMessage(this);
-      } catch (java.io.IOException e) {
-        throw new com.google.protobuf.InvalidProtocolBufferException(
-            e.getMessage()).setUnfinishedMessage(this);
-      } finally {
-        if (((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-          command_ = java.util.Collections.unmodifiableList(command_);
-        }
-        this.unknownFields = unknownFields.build();
-        makeExtensionsImmutable();
-      }
-    }
-    public static final com.google.protobuf.Descriptors.Descriptor
-        getDescriptor() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_SendCommandRequest_descriptor;
-    }
-
-    protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-        internalGetFieldAccessorTable() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_SendCommandRequest_fieldAccessorTable
-          .ensureFieldAccessorsInitialized(
-              org.yamcs.protobuf.Commanding.SendCommandRequest.class, org.yamcs.protobuf.Commanding.SendCommandRequest.Builder.class);
-    }
-
-    public static com.google.protobuf.Parser<SendCommandRequest> PARSER =
-        new com.google.protobuf.AbstractParser<SendCommandRequest>() {
-      public SendCommandRequest parsePartialFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return new SendCommandRequest(input, extensionRegistry);
-      }
-    };
-
-    @java.lang.Override
-    public com.google.protobuf.Parser<SendCommandRequest> getParserForType() {
-      return PARSER;
-    }
-
-    // repeated .commanding.CommandType command = 1;
-    public static final int COMMAND_FIELD_NUMBER = 1;
-    private java.util.List<org.yamcs.protobuf.Commanding.CommandType> command_;
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public java.util.List<org.yamcs.protobuf.Commanding.CommandType> getCommandList() {
-      return command_;
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public java.util.List<? extends org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-        getCommandOrBuilderList() {
-      return command_;
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public int getCommandCount() {
-      return command_.size();
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.CommandType getCommand(int index) {
-      return command_.get(index);
-    }
-    /**
-     * <code>repeated .commanding.CommandType command = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.CommandTypeOrBuilder getCommandOrBuilder(
-        int index) {
-      return command_.get(index);
-    }
-
-    private void initFields() {
-      command_ = java.util.Collections.emptyList();
-    }
-    private byte memoizedIsInitialized = -1;
-    public final boolean isInitialized() {
-      byte isInitialized = memoizedIsInitialized;
-      if (isInitialized != -1) return isInitialized == 1;
-
-      for (int i = 0; i < getCommandCount(); i++) {
-        if (!getCommand(i).isInitialized()) {
-          memoizedIsInitialized = 0;
-          return false;
-        }
-      }
-      memoizedIsInitialized = 1;
-      return true;
-    }
-
-    public void writeTo(com.google.protobuf.CodedOutputStream output)
-                        throws java.io.IOException {
-      getSerializedSize();
-      for (int i = 0; i < command_.size(); i++) {
-        output.writeMessage(1, command_.get(i));
-      }
-      getUnknownFields().writeTo(output);
-    }
-
-    private int memoizedSerializedSize = -1;
-    public int getSerializedSize() {
-      int size = memoizedSerializedSize;
-      if (size != -1) return size;
-
-      size = 0;
-      for (int i = 0; i < command_.size(); i++) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeMessageSize(1, command_.get(i));
-      }
-      size += getUnknownFields().getSerializedSize();
-      memoizedSerializedSize = size;
-      return size;
-    }
-
-    private static final long serialVersionUID = 0L;
-    @java.lang.Override
-    protected java.lang.Object writeReplace()
-        throws java.io.ObjectStreamException {
-      return super.writeReplace();
-    }
-
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(
-        com.google.protobuf.ByteString data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(
-        com.google.protobuf.ByteString data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(byte[] data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(
-        byte[] data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseDelimitedFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseDelimitedFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(
-        com.google.protobuf.CodedInputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.SendCommandRequest parseFrom(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-
-    public static Builder newBuilder() { return Builder.create(); }
-    public Builder newBuilderForType() { return newBuilder(); }
-    public static Builder newBuilder(org.yamcs.protobuf.Commanding.SendCommandRequest prototype) {
-      return newBuilder().mergeFrom(prototype);
-    }
-    public Builder toBuilder() { return newBuilder(this); }
-
-    @java.lang.Override
-    protected Builder newBuilderForType(
-        com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-      Builder builder = new Builder(parent);
-      return builder;
-    }
-    /**
-     * Protobuf type {@code commanding.SendCommandRequest}
-     */
-    public static final class Builder extends
-        com.google.protobuf.GeneratedMessage.Builder<Builder>
-       implements org.yamcs.protobuf.Commanding.SendCommandRequestOrBuilder {
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_SendCommandRequest_descriptor;
-      }
-
-      protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_SendCommandRequest_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                org.yamcs.protobuf.Commanding.SendCommandRequest.class, org.yamcs.protobuf.Commanding.SendCommandRequest.Builder.class);
-      }
-
-      // Construct using org.yamcs.protobuf.Commanding.SendCommandRequest.newBuilder()
-      private Builder() {
-        maybeForceBuilderInitialization();
-      }
-
-      private Builder(
-          com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-        super(parent);
-        maybeForceBuilderInitialization();
-      }
-      private void maybeForceBuilderInitialization() {
-        if (com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders) {
-          getCommandFieldBuilder();
-        }
-      }
-      private static Builder create() {
-        return new Builder();
-      }
-
-      public Builder clear() {
-        super.clear();
-        if (commandBuilder_ == null) {
-          command_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-        } else {
-          commandBuilder_.clear();
-        }
-        return this;
-      }
-
-      public Builder clone() {
-        return create().mergeFrom(buildPartial());
-      }
-
-      public com.google.protobuf.Descriptors.Descriptor
-          getDescriptorForType() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_SendCommandRequest_descriptor;
-      }
-
-      public org.yamcs.protobuf.Commanding.SendCommandRequest getDefaultInstanceForType() {
-        return org.yamcs.protobuf.Commanding.SendCommandRequest.getDefaultInstance();
-      }
-
-      public org.yamcs.protobuf.Commanding.SendCommandRequest build() {
-        org.yamcs.protobuf.Commanding.SendCommandRequest result = buildPartial();
-        if (!result.isInitialized()) {
-          throw newUninitializedMessageException(result);
-        }
-        return result;
-      }
-
-      public org.yamcs.protobuf.Commanding.SendCommandRequest buildPartial() {
-        org.yamcs.protobuf.Commanding.SendCommandRequest result = new org.yamcs.protobuf.Commanding.SendCommandRequest(this);
-        int from_bitField0_ = bitField0_;
-        if (commandBuilder_ == null) {
-          if (((bitField0_ & 0x00000001) == 0x00000001)) {
-            command_ = java.util.Collections.unmodifiableList(command_);
-            bitField0_ = (bitField0_ & ~0x00000001);
-          }
-          result.command_ = command_;
-        } else {
-          result.command_ = commandBuilder_.build();
-        }
-        onBuilt();
-        return result;
-      }
-
-      public Builder mergeFrom(com.google.protobuf.Message other) {
-        if (other instanceof org.yamcs.protobuf.Commanding.SendCommandRequest) {
-          return mergeFrom((org.yamcs.protobuf.Commanding.SendCommandRequest)other);
-        } else {
-          super.mergeFrom(other);
-          return this;
-        }
-      }
-
-      public Builder mergeFrom(org.yamcs.protobuf.Commanding.SendCommandRequest other) {
-        if (other == org.yamcs.protobuf.Commanding.SendCommandRequest.getDefaultInstance()) return this;
-        if (commandBuilder_ == null) {
-          if (!other.command_.isEmpty()) {
-            if (command_.isEmpty()) {
-              command_ = other.command_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-            } else {
-              ensureCommandIsMutable();
-              command_.addAll(other.command_);
-            }
-            onChanged();
-          }
-        } else {
-          if (!other.command_.isEmpty()) {
-            if (commandBuilder_.isEmpty()) {
-              commandBuilder_.dispose();
-              commandBuilder_ = null;
-              command_ = other.command_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-              commandBuilder_ = 
-                com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders ?
-                   getCommandFieldBuilder() : null;
-            } else {
-              commandBuilder_.addAllMessages(other.command_);
-            }
-          }
-        }
-        this.mergeUnknownFields(other.getUnknownFields());
-        return this;
-      }
-
-      public final boolean isInitialized() {
-        for (int i = 0; i < getCommandCount(); i++) {
-          if (!getCommand(i).isInitialized()) {
-            
-            return false;
-          }
-        }
-        return true;
-      }
-
-      public Builder mergeFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        org.yamcs.protobuf.Commanding.SendCommandRequest parsedMessage = null;
-        try {
-          parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          parsedMessage = (org.yamcs.protobuf.Commanding.SendCommandRequest) e.getUnfinishedMessage();
-          throw e;
-        } finally {
-          if (parsedMessage != null) {
-            mergeFrom(parsedMessage);
-          }
-        }
-        return this;
-      }
-      private int bitField0_;
-
-      // repeated .commanding.CommandType command = 1;
-      private java.util.List<org.yamcs.protobuf.Commanding.CommandType> command_ =
-        java.util.Collections.emptyList();
-      private void ensureCommandIsMutable() {
-        if (!((bitField0_ & 0x00000001) == 0x00000001)) {
-          command_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.CommandType>(command_);
-          bitField0_ |= 0x00000001;
-         }
-      }
-
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.CommandType, org.yamcs.protobuf.Commanding.CommandType.Builder, org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> commandBuilder_;
-
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.CommandType> getCommandList() {
-        if (commandBuilder_ == null) {
-          return java.util.Collections.unmodifiableList(command_);
-        } else {
-          return commandBuilder_.getMessageList();
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public int getCommandCount() {
-        if (commandBuilder_ == null) {
-          return command_.size();
-        } else {
-          return commandBuilder_.getCount();
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType getCommand(int index) {
-        if (commandBuilder_ == null) {
-          return command_.get(index);
-        } else {
-          return commandBuilder_.getMessage(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder setCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType value) {
-        if (commandBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandIsMutable();
-          command_.set(index, value);
-          onChanged();
-        } else {
-          commandBuilder_.setMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder setCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType.Builder builderForValue) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.set(index, builderForValue.build());
-          onChanged();
-        } else {
-          commandBuilder_.setMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(org.yamcs.protobuf.Commanding.CommandType value) {
-        if (commandBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandIsMutable();
-          command_.add(value);
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType value) {
-        if (commandBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureCommandIsMutable();
-          command_.add(index, value);
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(
-          org.yamcs.protobuf.Commanding.CommandType.Builder builderForValue) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.add(builderForValue.build());
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addCommand(
-          int index, org.yamcs.protobuf.Commanding.CommandType.Builder builderForValue) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.add(index, builderForValue.build());
-          onChanged();
-        } else {
-          commandBuilder_.addMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder addAllCommand(
-          java.lang.Iterable<? extends org.yamcs.protobuf.Commanding.CommandType> values) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          super.addAll(values, command_);
-          onChanged();
-        } else {
-          commandBuilder_.addAllMessages(values);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder clearCommand() {
-        if (commandBuilder_ == null) {
-          command_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-          onChanged();
-        } else {
-          commandBuilder_.clear();
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public Builder removeCommand(int index) {
-        if (commandBuilder_ == null) {
-          ensureCommandIsMutable();
-          command_.remove(index);
-          onChanged();
-        } else {
-          commandBuilder_.remove(index);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType.Builder getCommandBuilder(
-          int index) {
-        return getCommandFieldBuilder().getBuilder(index);
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandTypeOrBuilder getCommandOrBuilder(
-          int index) {
-        if (commandBuilder_ == null) {
-          return command_.get(index);  } else {
-          return commandBuilder_.getMessageOrBuilder(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public java.util.List<? extends org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-           getCommandOrBuilderList() {
-        if (commandBuilder_ != null) {
-          return commandBuilder_.getMessageOrBuilderList();
-        } else {
-          return java.util.Collections.unmodifiableList(command_);
-        }
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType.Builder addCommandBuilder() {
-        return getCommandFieldBuilder().addBuilder(
-            org.yamcs.protobuf.Commanding.CommandType.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.CommandType.Builder addCommandBuilder(
-          int index) {
-        return getCommandFieldBuilder().addBuilder(
-            index, org.yamcs.protobuf.Commanding.CommandType.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.CommandType command = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.CommandType.Builder> 
-           getCommandBuilderList() {
-        return getCommandFieldBuilder().getBuilderList();
-      }
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.CommandType, org.yamcs.protobuf.Commanding.CommandType.Builder, org.yamcs.protobuf.Commanding.CommandTypeOrBuilder> 
-          getCommandFieldBuilder() {
-        if (commandBuilder_ == null) {
-          commandBuilder_ = new com.google.protobuf.RepeatedFieldBuilder<
-              org.yamcs.protobuf.Commanding.CommandType, org.yamcs.protobuf.Commanding.CommandType.Builder, org.yamcs.protobuf.Commanding.CommandTypeOrBuilder>(
-                  command_,
-                  ((bitField0_ & 0x00000001) == 0x00000001),
-                  getParentForChildren(),
-                  isClean());
-          command_ = null;
-        }
-        return commandBuilder_;
-      }
-
-      // @@protoc_insertion_point(builder_scope:commanding.SendCommandRequest)
-    }
-
-    static {
-      defaultInstance = new SendCommandRequest(true);
-      defaultInstance.initFields();
-    }
-
-    // @@protoc_insertion_point(class_scope:commanding.SendCommandRequest)
-  }
-
-  public interface BulkGetCommandResponseOrBuilder
-      extends com.google.protobuf.MessageOrBuilder {
-
-    // repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    java.util.List<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> 
-        getResponseList();
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse getResponse(int index);
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    int getResponseCount();
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    java.util.List<? extends org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder> 
-        getResponseOrBuilderList();
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder getResponseOrBuilder(
-        int index);
-  }
-  /**
-   * Protobuf type {@code commanding.BulkGetCommandResponse}
-   */
-  public static final class BulkGetCommandResponse extends
-      com.google.protobuf.GeneratedMessage
-      implements BulkGetCommandResponseOrBuilder {
-    // Use BulkGetCommandResponse.newBuilder() to construct.
-    private BulkGetCommandResponse(com.google.protobuf.GeneratedMessage.Builder<?> builder) {
-      super(builder);
-      this.unknownFields = builder.getUnknownFields();
-    }
-    private BulkGetCommandResponse(boolean noInit) { this.unknownFields = com.google.protobuf.UnknownFieldSet.getDefaultInstance(); }
-
-    private static final BulkGetCommandResponse defaultInstance;
-    public static BulkGetCommandResponse getDefaultInstance() {
-      return defaultInstance;
-    }
-
-    public BulkGetCommandResponse getDefaultInstanceForType() {
-      return defaultInstance;
-    }
-
-    private final com.google.protobuf.UnknownFieldSet unknownFields;
-    @java.lang.Override
-    public final com.google.protobuf.UnknownFieldSet
-        getUnknownFields() {
-      return this.unknownFields;
-    }
-    private BulkGetCommandResponse(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      initFields();
-      int mutable_bitField0_ = 0;
-      com.google.protobuf.UnknownFieldSet.Builder unknownFields =
-          com.google.protobuf.UnknownFieldSet.newBuilder();
-      try {
-        boolean done = false;
-        while (!done) {
-          int tag = input.readTag();
-          switch (tag) {
-            case 0:
-              done = true;
-              break;
-            default: {
-              if (!parseUnknownField(input, unknownFields,
-                                     extensionRegistry, tag)) {
-                done = true;
-              }
-              break;
-            }
-            case 10: {
-              if (!((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-                response_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse>();
-                mutable_bitField0_ |= 0x00000001;
-              }
-              response_.add(input.readMessage(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.PARSER, extensionRegistry));
-              break;
-            }
-          }
-        }
-      } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-        throw e.setUnfinishedMessage(this);
-      } catch (java.io.IOException e) {
-        throw new com.google.protobuf.InvalidProtocolBufferException(
-            e.getMessage()).setUnfinishedMessage(this);
-      } finally {
-        if (((mutable_bitField0_ & 0x00000001) == 0x00000001)) {
-          response_ = java.util.Collections.unmodifiableList(response_);
-        }
-        this.unknownFields = unknownFields.build();
-        makeExtensionsImmutable();
-      }
-    }
-    public static final com.google.protobuf.Descriptors.Descriptor
-        getDescriptor() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_descriptor;
-    }
-
-    protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-        internalGetFieldAccessorTable() {
-      return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_fieldAccessorTable
-          .ensureFieldAccessorsInitialized(
-              org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder.class);
-    }
-
-    public static com.google.protobuf.Parser<BulkGetCommandResponse> PARSER =
-        new com.google.protobuf.AbstractParser<BulkGetCommandResponse>() {
-      public BulkGetCommandResponse parsePartialFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return new BulkGetCommandResponse(input, extensionRegistry);
-      }
-    };
-
-    @java.lang.Override
-    public com.google.protobuf.Parser<BulkGetCommandResponse> getParserForType() {
-      return PARSER;
-    }
-
-    public interface CommandInfoResponseOrBuilder
-        extends com.google.protobuf.MessageOrBuilder {
-
-      // optional .yamcs.NamedObjectId id = 1;
-      /**
-       * <code>optional .yamcs.NamedObjectId id = 1;</code>
-       */
-      boolean hasId();
-      /**
-       * <code>optional .yamcs.NamedObjectId id = 1;</code>
-       */
-      org.yamcs.protobuf.Yamcs.NamedObjectId getId();
-      /**
-       * <code>optional .yamcs.NamedObjectId id = 1;</code>
-       */
-      org.yamcs.protobuf.Yamcs.NamedObjectIdOrBuilder getIdOrBuilder();
-
-      // optional .mdb.CommandInfo command = 2;
-      /**
-       * <code>optional .mdb.CommandInfo command = 2;</code>
-       */
-      boolean hasCommand();
-      /**
-       * <code>optional .mdb.CommandInfo command = 2;</code>
-       */
-      org.yamcs.protobuf.Mdb.CommandInfo getCommand();
-      /**
-       * <code>optional .mdb.CommandInfo command = 2;</code>
-       */
-      org.yamcs.protobuf.Mdb.CommandInfoOrBuilder getCommandOrBuilder();
-    }
-    /**
-     * Protobuf type {@code commanding.BulkGetCommandResponse.CommandInfoResponse}
-     */
-    public static final class CommandInfoResponse extends
-        com.google.protobuf.GeneratedMessage
-        implements CommandInfoResponseOrBuilder {
-      // Use CommandInfoResponse.newBuilder() to construct.
-      private CommandInfoResponse(com.google.protobuf.GeneratedMessage.Builder<?> builder) {
-        super(builder);
-        this.unknownFields = builder.getUnknownFields();
-      }
-      private CommandInfoResponse(boolean noInit) { this.unknownFields = com.google.protobuf.UnknownFieldSet.getDefaultInstance(); }
-
-      private static final CommandInfoResponse defaultInstance;
-      public static CommandInfoResponse getDefaultInstance() {
-        return defaultInstance;
-      }
-
-      public CommandInfoResponse getDefaultInstanceForType() {
-        return defaultInstance;
-      }
-
-      private final com.google.protobuf.UnknownFieldSet unknownFields;
-      @java.lang.Override
-      public final com.google.protobuf.UnknownFieldSet
-          getUnknownFields() {
-        return this.unknownFields;
-      }
-      private CommandInfoResponse(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        initFields();
-        int mutable_bitField0_ = 0;
-        com.google.protobuf.UnknownFieldSet.Builder unknownFields =
-            com.google.protobuf.UnknownFieldSet.newBuilder();
-        try {
-          boolean done = false;
-          while (!done) {
-            int tag = input.readTag();
-            switch (tag) {
-              case 0:
-                done = true;
-                break;
-              default: {
-                if (!parseUnknownField(input, unknownFields,
-                                       extensionRegistry, tag)) {
-                  done = true;
-                }
-                break;
-              }
-              case 10: {
-                org.yamcs.protobuf.Yamcs.NamedObjectId.Builder subBuilder = null;
-                if (((bitField0_ & 0x00000001) == 0x00000001)) {
-                  subBuilder = id_.toBuilder();
-                }
-                id_ = input.readMessage(org.yamcs.protobuf.Yamcs.NamedObjectId.PARSER, extensionRegistry);
-                if (subBuilder != null) {
-                  subBuilder.mergeFrom(id_);
-                  id_ = subBuilder.buildPartial();
-                }
-                bitField0_ |= 0x00000001;
-                break;
-              }
-              case 18: {
-                org.yamcs.protobuf.Mdb.CommandInfo.Builder subBuilder = null;
-                if (((bitField0_ & 0x00000002) == 0x00000002)) {
-                  subBuilder = command_.toBuilder();
-                }
-                command_ = input.readMessage(org.yamcs.protobuf.Mdb.CommandInfo.PARSER, extensionRegistry);
-                if (subBuilder != null) {
-                  subBuilder.mergeFrom(command_);
-                  command_ = subBuilder.buildPartial();
-                }
-                bitField0_ |= 0x00000002;
-                break;
-              }
-            }
-          }
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          throw e.setUnfinishedMessage(this);
-        } catch (java.io.IOException e) {
-          throw new com.google.protobuf.InvalidProtocolBufferException(
-              e.getMessage()).setUnfinishedMessage(this);
-        } finally {
-          this.unknownFields = unknownFields.build();
-          makeExtensionsImmutable();
-        }
-      }
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_descriptor;
-      }
-
-      protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder.class);
-      }
-
-      public static com.google.protobuf.Parser<CommandInfoResponse> PARSER =
-          new com.google.protobuf.AbstractParser<CommandInfoResponse>() {
-        public CommandInfoResponse parsePartialFrom(
-            com.google.protobuf.CodedInputStream input,
-            com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-            throws com.google.protobuf.InvalidProtocolBufferException {
-          return new CommandInfoResponse(input, extensionRegistry);
-        }
-      };
-
-      @java.lang.Override
-      public com.google.protobuf.Parser<CommandInfoResponse> getParserForType() {
-        return PARSER;
-      }
-
-      private int bitField0_;
-      // optional .yamcs.NamedObjectId id = 1;
-      public static final int ID_FIELD_NUMBER = 1;
-      private org.yamcs.protobuf.Yamcs.NamedObjectId id_;
-      /**
-       * <code>optional .yamcs.NamedObjectId id = 1;</code>
-       */
-      public boolean hasId() {
-        return ((bitField0_ & 0x00000001) == 0x00000001);
-      }
-      /**
-       * <code>optional .yamcs.NamedObjectId id = 1;</code>
-       */
-      public org.yamcs.protobuf.Yamcs.NamedObjectId getId() {
-        return id_;
-      }
-      /**
-       * <code>optional .yamcs.NamedObjectId id = 1;</code>
-       */
-      public org.yamcs.protobuf.Yamcs.NamedObjectIdOrBuilder getIdOrBuilder() {
-        return id_;
-      }
-
-      // optional .mdb.CommandInfo command = 2;
-      public static final int COMMAND_FIELD_NUMBER = 2;
-      private org.yamcs.protobuf.Mdb.CommandInfo command_;
-      /**
-       * <code>optional .mdb.CommandInfo command = 2;</code>
-       */
-      public boolean hasCommand() {
-        return ((bitField0_ & 0x00000002) == 0x00000002);
-      }
-      /**
-       * <code>optional .mdb.CommandInfo command = 2;</code>
-       */
-      public org.yamcs.protobuf.Mdb.CommandInfo getCommand() {
-        return command_;
-      }
-      /**
-       * <code>optional .mdb.CommandInfo command = 2;</code>
-       */
-      public org.yamcs.protobuf.Mdb.CommandInfoOrBuilder getCommandOrBuilder() {
-        return command_;
-      }
-
-      private void initFields() {
-        id_ = org.yamcs.protobuf.Yamcs.NamedObjectId.getDefaultInstance();
-        command_ = org.yamcs.protobuf.Mdb.CommandInfo.getDefaultInstance();
-      }
-      private byte memoizedIsInitialized = -1;
-      public final boolean isInitialized() {
-        byte isInitialized = memoizedIsInitialized;
-        if (isInitialized != -1) return isInitialized == 1;
-
-        if (hasId()) {
-          if (!getId().isInitialized()) {
-            memoizedIsInitialized = 0;
-            return false;
-          }
-        }
-        if (hasCommand()) {
-          if (!getCommand().isInitialized()) {
-            memoizedIsInitialized = 0;
-            return false;
-          }
-        }
-        memoizedIsInitialized = 1;
-        return true;
-      }
-
-      public void writeTo(com.google.protobuf.CodedOutputStream output)
-                          throws java.io.IOException {
-        getSerializedSize();
-        if (((bitField0_ & 0x00000001) == 0x00000001)) {
-          output.writeMessage(1, id_);
-        }
-        if (((bitField0_ & 0x00000002) == 0x00000002)) {
-          output.writeMessage(2, command_);
-        }
-        getUnknownFields().writeTo(output);
-      }
-
-      private int memoizedSerializedSize = -1;
-      public int getSerializedSize() {
-        int size = memoizedSerializedSize;
-        if (size != -1) return size;
-
-        size = 0;
-        if (((bitField0_ & 0x00000001) == 0x00000001)) {
-          size += com.google.protobuf.CodedOutputStream
-            .computeMessageSize(1, id_);
-        }
-        if (((bitField0_ & 0x00000002) == 0x00000002)) {
-          size += com.google.protobuf.CodedOutputStream
-            .computeMessageSize(2, command_);
-        }
-        size += getUnknownFields().getSerializedSize();
-        memoizedSerializedSize = size;
-        return size;
-      }
-
-      private static final long serialVersionUID = 0L;
-      @java.lang.Override
-      protected java.lang.Object writeReplace()
-          throws java.io.ObjectStreamException {
-        return super.writeReplace();
-      }
-
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(
-          com.google.protobuf.ByteString data)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return PARSER.parseFrom(data);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(
-          com.google.protobuf.ByteString data,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return PARSER.parseFrom(data, extensionRegistry);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(byte[] data)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return PARSER.parseFrom(data);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(
-          byte[] data,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        return PARSER.parseFrom(data, extensionRegistry);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(java.io.InputStream input)
-          throws java.io.IOException {
-        return PARSER.parseFrom(input);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(
-          java.io.InputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        return PARSER.parseFrom(input, extensionRegistry);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseDelimitedFrom(java.io.InputStream input)
-          throws java.io.IOException {
-        return PARSER.parseDelimitedFrom(input);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseDelimitedFrom(
-          java.io.InputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        return PARSER.parseDelimitedFrom(input, extensionRegistry);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(
-          com.google.protobuf.CodedInputStream input)
-          throws java.io.IOException {
-        return PARSER.parseFrom(input);
-      }
-      public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parseFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        return PARSER.parseFrom(input, extensionRegistry);
-      }
-
-      public static Builder newBuilder() { return Builder.create(); }
-      public Builder newBuilderForType() { return newBuilder(); }
-      public static Builder newBuilder(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse prototype) {
-        return newBuilder().mergeFrom(prototype);
-      }
-      public Builder toBuilder() { return newBuilder(this); }
-
-      @java.lang.Override
-      protected Builder newBuilderForType(
-          com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-        Builder builder = new Builder(parent);
-        return builder;
-      }
-      /**
-       * Protobuf type {@code commanding.BulkGetCommandResponse.CommandInfoResponse}
-       */
-      public static final class Builder extends
-          com.google.protobuf.GeneratedMessage.Builder<Builder>
-         implements org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder {
-        public static final com.google.protobuf.Descriptors.Descriptor
-            getDescriptor() {
-          return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_descriptor;
-        }
-
-        protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-            internalGetFieldAccessorTable() {
-          return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_fieldAccessorTable
-              .ensureFieldAccessorsInitialized(
-                  org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder.class);
-        }
-
-        // Construct using org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.newBuilder()
-        private Builder() {
-          maybeForceBuilderInitialization();
-        }
-
-        private Builder(
-            com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-          super(parent);
-          maybeForceBuilderInitialization();
-        }
-        private void maybeForceBuilderInitialization() {
-          if (com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders) {
-            getIdFieldBuilder();
-            getCommandFieldBuilder();
-          }
-        }
-        private static Builder create() {
-          return new Builder();
-        }
-
-        public Builder clear() {
-          super.clear();
-          if (idBuilder_ == null) {
-            id_ = org.yamcs.protobuf.Yamcs.NamedObjectId.getDefaultInstance();
-          } else {
-            idBuilder_.clear();
-          }
-          bitField0_ = (bitField0_ & ~0x00000001);
-          if (commandBuilder_ == null) {
-            command_ = org.yamcs.protobuf.Mdb.CommandInfo.getDefaultInstance();
-          } else {
-            commandBuilder_.clear();
-          }
-          bitField0_ = (bitField0_ & ~0x00000002);
-          return this;
-        }
-
-        public Builder clone() {
-          return create().mergeFrom(buildPartial());
-        }
-
-        public com.google.protobuf.Descriptors.Descriptor
-            getDescriptorForType() {
-          return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_descriptor;
-        }
-
-        public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse getDefaultInstanceForType() {
-          return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.getDefaultInstance();
-        }
-
-        public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse build() {
-          org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse result = buildPartial();
-          if (!result.isInitialized()) {
-            throw newUninitializedMessageException(result);
-          }
-          return result;
-        }
-
-        public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse buildPartial() {
-          org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse result = new org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse(this);
-          int from_bitField0_ = bitField0_;
-          int to_bitField0_ = 0;
-          if (((from_bitField0_ & 0x00000001) == 0x00000001)) {
-            to_bitField0_ |= 0x00000001;
-          }
-          if (idBuilder_ == null) {
-            result.id_ = id_;
-          } else {
-            result.id_ = idBuilder_.build();
-          }
-          if (((from_bitField0_ & 0x00000002) == 0x00000002)) {
-            to_bitField0_ |= 0x00000002;
-          }
-          if (commandBuilder_ == null) {
-            result.command_ = command_;
-          } else {
-            result.command_ = commandBuilder_.build();
-          }
-          result.bitField0_ = to_bitField0_;
-          onBuilt();
-          return result;
-        }
-
-        public Builder mergeFrom(com.google.protobuf.Message other) {
-          if (other instanceof org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse) {
-            return mergeFrom((org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse)other);
-          } else {
-            super.mergeFrom(other);
-            return this;
-          }
-        }
-
-        public Builder mergeFrom(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse other) {
-          if (other == org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.getDefaultInstance()) return this;
-          if (other.hasId()) {
-            mergeId(other.getId());
-          }
-          if (other.hasCommand()) {
-            mergeCommand(other.getCommand());
-          }
-          this.mergeUnknownFields(other.getUnknownFields());
-          return this;
-        }
-
-        public final boolean isInitialized() {
-          if (hasId()) {
-            if (!getId().isInitialized()) {
-              
-              return false;
-            }
-          }
-          if (hasCommand()) {
-            if (!getCommand().isInitialized()) {
-              
-              return false;
-            }
-          }
-          return true;
-        }
-
-        public Builder mergeFrom(
-            com.google.protobuf.CodedInputStream input,
-            com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-            throws java.io.IOException {
-          org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse parsedMessage = null;
-          try {
-            parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
-          } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-            parsedMessage = (org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse) e.getUnfinishedMessage();
-            throw e;
-          } finally {
-            if (parsedMessage != null) {
-              mergeFrom(parsedMessage);
-            }
-          }
-          return this;
-        }
-        private int bitField0_;
-
-        // optional .yamcs.NamedObjectId id = 1;
-        private org.yamcs.protobuf.Yamcs.NamedObjectId id_ = org.yamcs.protobuf.Yamcs.NamedObjectId.getDefaultInstance();
-        private com.google.protobuf.SingleFieldBuilder<
-            org.yamcs.protobuf.Yamcs.NamedObjectId, org.yamcs.protobuf.Yamcs.NamedObjectId.Builder, org.yamcs.protobuf.Yamcs.NamedObjectIdOrBuilder> idBuilder_;
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public boolean hasId() {
-          return ((bitField0_ & 0x00000001) == 0x00000001);
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public org.yamcs.protobuf.Yamcs.NamedObjectId getId() {
-          if (idBuilder_ == null) {
-            return id_;
-          } else {
-            return idBuilder_.getMessage();
-          }
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public Builder setId(org.yamcs.protobuf.Yamcs.NamedObjectId value) {
-          if (idBuilder_ == null) {
-            if (value == null) {
-              throw new NullPointerException();
-            }
-            id_ = value;
-            onChanged();
-          } else {
-            idBuilder_.setMessage(value);
-          }
-          bitField0_ |= 0x00000001;
-          return this;
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public Builder setId(
-            org.yamcs.protobuf.Yamcs.NamedObjectId.Builder builderForValue) {
-          if (idBuilder_ == null) {
-            id_ = builderForValue.build();
-            onChanged();
-          } else {
-            idBuilder_.setMessage(builderForValue.build());
-          }
-          bitField0_ |= 0x00000001;
-          return this;
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public Builder mergeId(org.yamcs.protobuf.Yamcs.NamedObjectId value) {
-          if (idBuilder_ == null) {
-            if (((bitField0_ & 0x00000001) == 0x00000001) &&
-                id_ != org.yamcs.protobuf.Yamcs.NamedObjectId.getDefaultInstance()) {
-              id_ =
-                org.yamcs.protobuf.Yamcs.NamedObjectId.newBuilder(id_).mergeFrom(value).buildPartial();
-            } else {
-              id_ = value;
-            }
-            onChanged();
-          } else {
-            idBuilder_.mergeFrom(value);
-          }
-          bitField0_ |= 0x00000001;
-          return this;
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public Builder clearId() {
-          if (idBuilder_ == null) {
-            id_ = org.yamcs.protobuf.Yamcs.NamedObjectId.getDefaultInstance();
-            onChanged();
-          } else {
-            idBuilder_.clear();
-          }
-          bitField0_ = (bitField0_ & ~0x00000001);
-          return this;
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public org.yamcs.protobuf.Yamcs.NamedObjectId.Builder getIdBuilder() {
-          bitField0_ |= 0x00000001;
-          onChanged();
-          return getIdFieldBuilder().getBuilder();
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        public org.yamcs.protobuf.Yamcs.NamedObjectIdOrBuilder getIdOrBuilder() {
-          if (idBuilder_ != null) {
-            return idBuilder_.getMessageOrBuilder();
-          } else {
-            return id_;
-          }
-        }
-        /**
-         * <code>optional .yamcs.NamedObjectId id = 1;</code>
-         */
-        private com.google.protobuf.SingleFieldBuilder<
-            org.yamcs.protobuf.Yamcs.NamedObjectId, org.yamcs.protobuf.Yamcs.NamedObjectId.Builder, org.yamcs.protobuf.Yamcs.NamedObjectIdOrBuilder> 
-            getIdFieldBuilder() {
-          if (idBuilder_ == null) {
-            idBuilder_ = new com.google.protobuf.SingleFieldBuilder<
-                org.yamcs.protobuf.Yamcs.NamedObjectId, org.yamcs.protobuf.Yamcs.NamedObjectId.Builder, org.yamcs.protobuf.Yamcs.NamedObjectIdOrBuilder>(
-                    id_,
-                    getParentForChildren(),
-                    isClean());
-            id_ = null;
-          }
-          return idBuilder_;
-        }
-
-        // optional .mdb.CommandInfo command = 2;
-        private org.yamcs.protobuf.Mdb.CommandInfo command_ = org.yamcs.protobuf.Mdb.CommandInfo.getDefaultInstance();
-        private com.google.protobuf.SingleFieldBuilder<
-            org.yamcs.protobuf.Mdb.CommandInfo, org.yamcs.protobuf.Mdb.CommandInfo.Builder, org.yamcs.protobuf.Mdb.CommandInfoOrBuilder> commandBuilder_;
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public boolean hasCommand() {
-          return ((bitField0_ & 0x00000002) == 0x00000002);
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public org.yamcs.protobuf.Mdb.CommandInfo getCommand() {
-          if (commandBuilder_ == null) {
-            return command_;
-          } else {
-            return commandBuilder_.getMessage();
-          }
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public Builder setCommand(org.yamcs.protobuf.Mdb.CommandInfo value) {
-          if (commandBuilder_ == null) {
-            if (value == null) {
-              throw new NullPointerException();
-            }
-            command_ = value;
-            onChanged();
-          } else {
-            commandBuilder_.setMessage(value);
-          }
-          bitField0_ |= 0x00000002;
-          return this;
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public Builder setCommand(
-            org.yamcs.protobuf.Mdb.CommandInfo.Builder builderForValue) {
-          if (commandBuilder_ == null) {
-            command_ = builderForValue.build();
-            onChanged();
-          } else {
-            commandBuilder_.setMessage(builderForValue.build());
-          }
-          bitField0_ |= 0x00000002;
-          return this;
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public Builder mergeCommand(org.yamcs.protobuf.Mdb.CommandInfo value) {
-          if (commandBuilder_ == null) {
-            if (((bitField0_ & 0x00000002) == 0x00000002) &&
-                command_ != org.yamcs.protobuf.Mdb.CommandInfo.getDefaultInstance()) {
-              command_ =
-                org.yamcs.protobuf.Mdb.CommandInfo.newBuilder(command_).mergeFrom(value).buildPartial();
-            } else {
-              command_ = value;
-            }
-            onChanged();
-          } else {
-            commandBuilder_.mergeFrom(value);
-          }
-          bitField0_ |= 0x00000002;
-          return this;
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public Builder clearCommand() {
-          if (commandBuilder_ == null) {
-            command_ = org.yamcs.protobuf.Mdb.CommandInfo.getDefaultInstance();
-            onChanged();
-          } else {
-            commandBuilder_.clear();
-          }
-          bitField0_ = (bitField0_ & ~0x00000002);
-          return this;
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public org.yamcs.protobuf.Mdb.CommandInfo.Builder getCommandBuilder() {
-          bitField0_ |= 0x00000002;
-          onChanged();
-          return getCommandFieldBuilder().getBuilder();
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        public org.yamcs.protobuf.Mdb.CommandInfoOrBuilder getCommandOrBuilder() {
-          if (commandBuilder_ != null) {
-            return commandBuilder_.getMessageOrBuilder();
-          } else {
-            return command_;
-          }
-        }
-        /**
-         * <code>optional .mdb.CommandInfo command = 2;</code>
-         */
-        private com.google.protobuf.SingleFieldBuilder<
-            org.yamcs.protobuf.Mdb.CommandInfo, org.yamcs.protobuf.Mdb.CommandInfo.Builder, org.yamcs.protobuf.Mdb.CommandInfoOrBuilder> 
-            getCommandFieldBuilder() {
-          if (commandBuilder_ == null) {
-            commandBuilder_ = new com.google.protobuf.SingleFieldBuilder<
-                org.yamcs.protobuf.Mdb.CommandInfo, org.yamcs.protobuf.Mdb.CommandInfo.Builder, org.yamcs.protobuf.Mdb.CommandInfoOrBuilder>(
-                    command_,
-                    getParentForChildren(),
-                    isClean());
-            command_ = null;
-          }
-          return commandBuilder_;
-        }
-
-        // @@protoc_insertion_point(builder_scope:commanding.BulkGetCommandResponse.CommandInfoResponse)
-      }
-
-      static {
-        defaultInstance = new CommandInfoResponse(true);
-        defaultInstance.initFields();
-      }
-
-      // @@protoc_insertion_point(class_scope:commanding.BulkGetCommandResponse.CommandInfoResponse)
-    }
-
-    // repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;
-    public static final int RESPONSE_FIELD_NUMBER = 1;
-    private java.util.List<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> response_;
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    public java.util.List<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> getResponseList() {
-      return response_;
-    }
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    public java.util.List<? extends org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder> 
-        getResponseOrBuilderList() {
-      return response_;
-    }
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    public int getResponseCount() {
-      return response_.size();
-    }
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse getResponse(int index) {
-      return response_.get(index);
-    }
-    /**
-     * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-     */
-    public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder getResponseOrBuilder(
-        int index) {
-      return response_.get(index);
-    }
-
-    private void initFields() {
-      response_ = java.util.Collections.emptyList();
-    }
-    private byte memoizedIsInitialized = -1;
-    public final boolean isInitialized() {
-      byte isInitialized = memoizedIsInitialized;
-      if (isInitialized != -1) return isInitialized == 1;
-
-      for (int i = 0; i < getResponseCount(); i++) {
-        if (!getResponse(i).isInitialized()) {
-          memoizedIsInitialized = 0;
-          return false;
-        }
-      }
-      memoizedIsInitialized = 1;
-      return true;
-    }
-
-    public void writeTo(com.google.protobuf.CodedOutputStream output)
-                        throws java.io.IOException {
-      getSerializedSize();
-      for (int i = 0; i < response_.size(); i++) {
-        output.writeMessage(1, response_.get(i));
-      }
-      getUnknownFields().writeTo(output);
-    }
-
-    private int memoizedSerializedSize = -1;
-    public int getSerializedSize() {
-      int size = memoizedSerializedSize;
-      if (size != -1) return size;
-
-      size = 0;
-      for (int i = 0; i < response_.size(); i++) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeMessageSize(1, response_.get(i));
-      }
-      size += getUnknownFields().getSerializedSize();
-      memoizedSerializedSize = size;
-      return size;
-    }
-
-    private static final long serialVersionUID = 0L;
-    @java.lang.Override
-    protected java.lang.Object writeReplace()
-        throws java.io.ObjectStreamException {
-      return super.writeReplace();
-    }
-
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(
-        com.google.protobuf.ByteString data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(
-        com.google.protobuf.ByteString data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(byte[] data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(
-        byte[] data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseDelimitedFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseDelimitedFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseDelimitedFrom(input, extensionRegistry);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(
-        com.google.protobuf.CodedInputStream input)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input);
-    }
-    public static org.yamcs.protobuf.Commanding.BulkGetCommandResponse parseFrom(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return PARSER.parseFrom(input, extensionRegistry);
-    }
-
-    public static Builder newBuilder() { return Builder.create(); }
-    public Builder newBuilderForType() { return newBuilder(); }
-    public static Builder newBuilder(org.yamcs.protobuf.Commanding.BulkGetCommandResponse prototype) {
-      return newBuilder().mergeFrom(prototype);
-    }
-    public Builder toBuilder() { return newBuilder(this); }
-
-    @java.lang.Override
-    protected Builder newBuilderForType(
-        com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-      Builder builder = new Builder(parent);
-      return builder;
-    }
-    /**
-     * Protobuf type {@code commanding.BulkGetCommandResponse}
-     */
-    public static final class Builder extends
-        com.google.protobuf.GeneratedMessage.Builder<Builder>
-       implements org.yamcs.protobuf.Commanding.BulkGetCommandResponseOrBuilder {
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_descriptor;
-      }
-
-      protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder.class);
-      }
-
-      // Construct using org.yamcs.protobuf.Commanding.BulkGetCommandResponse.newBuilder()
-      private Builder() {
-        maybeForceBuilderInitialization();
-      }
-
-      private Builder(
-          com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-        super(parent);
-        maybeForceBuilderInitialization();
-      }
-      private void maybeForceBuilderInitialization() {
-        if (com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders) {
-          getResponseFieldBuilder();
-        }
-      }
-      private static Builder create() {
-        return new Builder();
-      }
-
-      public Builder clear() {
-        super.clear();
-        if (responseBuilder_ == null) {
-          response_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-        } else {
-          responseBuilder_.clear();
-        }
-        return this;
-      }
-
-      public Builder clone() {
-        return create().mergeFrom(buildPartial());
-      }
-
-      public com.google.protobuf.Descriptors.Descriptor
-          getDescriptorForType() {
-        return org.yamcs.protobuf.Commanding.internal_static_commanding_BulkGetCommandResponse_descriptor;
-      }
-
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse getDefaultInstanceForType() {
-        return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.getDefaultInstance();
-      }
-
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse build() {
-        org.yamcs.protobuf.Commanding.BulkGetCommandResponse result = buildPartial();
-        if (!result.isInitialized()) {
-          throw newUninitializedMessageException(result);
-        }
-        return result;
-      }
-
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse buildPartial() {
-        org.yamcs.protobuf.Commanding.BulkGetCommandResponse result = new org.yamcs.protobuf.Commanding.BulkGetCommandResponse(this);
-        int from_bitField0_ = bitField0_;
-        if (responseBuilder_ == null) {
-          if (((bitField0_ & 0x00000001) == 0x00000001)) {
-            response_ = java.util.Collections.unmodifiableList(response_);
-            bitField0_ = (bitField0_ & ~0x00000001);
-          }
-          result.response_ = response_;
-        } else {
-          result.response_ = responseBuilder_.build();
-        }
-        onBuilt();
-        return result;
-      }
-
-      public Builder mergeFrom(com.google.protobuf.Message other) {
-        if (other instanceof org.yamcs.protobuf.Commanding.BulkGetCommandResponse) {
-          return mergeFrom((org.yamcs.protobuf.Commanding.BulkGetCommandResponse)other);
-        } else {
-          super.mergeFrom(other);
-          return this;
-        }
-      }
-
-      public Builder mergeFrom(org.yamcs.protobuf.Commanding.BulkGetCommandResponse other) {
-        if (other == org.yamcs.protobuf.Commanding.BulkGetCommandResponse.getDefaultInstance()) return this;
-        if (responseBuilder_ == null) {
-          if (!other.response_.isEmpty()) {
-            if (response_.isEmpty()) {
-              response_ = other.response_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-            } else {
-              ensureResponseIsMutable();
-              response_.addAll(other.response_);
-            }
-            onChanged();
-          }
-        } else {
-          if (!other.response_.isEmpty()) {
-            if (responseBuilder_.isEmpty()) {
-              responseBuilder_.dispose();
-              responseBuilder_ = null;
-              response_ = other.response_;
-              bitField0_ = (bitField0_ & ~0x00000001);
-              responseBuilder_ = 
-                com.google.protobuf.GeneratedMessage.alwaysUseFieldBuilders ?
-                   getResponseFieldBuilder() : null;
-            } else {
-              responseBuilder_.addAllMessages(other.response_);
-            }
-          }
-        }
-        this.mergeUnknownFields(other.getUnknownFields());
-        return this;
-      }
-
-      public final boolean isInitialized() {
-        for (int i = 0; i < getResponseCount(); i++) {
-          if (!getResponse(i).isInitialized()) {
-            
-            return false;
-          }
-        }
-        return true;
-      }
-
-      public Builder mergeFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        org.yamcs.protobuf.Commanding.BulkGetCommandResponse parsedMessage = null;
-        try {
-          parsedMessage = PARSER.parsePartialFrom(input, extensionRegistry);
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          parsedMessage = (org.yamcs.protobuf.Commanding.BulkGetCommandResponse) e.getUnfinishedMessage();
-          throw e;
-        } finally {
-          if (parsedMessage != null) {
-            mergeFrom(parsedMessage);
-          }
-        }
-        return this;
-      }
-      private int bitField0_;
-
-      // repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;
-      private java.util.List<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> response_ =
-        java.util.Collections.emptyList();
-      private void ensureResponseIsMutable() {
-        if (!((bitField0_ & 0x00000001) == 0x00000001)) {
-          response_ = new java.util.ArrayList<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse>(response_);
-          bitField0_ |= 0x00000001;
-         }
-      }
-
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder> responseBuilder_;
-
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> getResponseList() {
-        if (responseBuilder_ == null) {
-          return java.util.Collections.unmodifiableList(response_);
-        } else {
-          return responseBuilder_.getMessageList();
-        }
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public int getResponseCount() {
-        if (responseBuilder_ == null) {
-          return response_.size();
-        } else {
-          return responseBuilder_.getCount();
-        }
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse getResponse(int index) {
-        if (responseBuilder_ == null) {
-          return response_.get(index);
-        } else {
-          return responseBuilder_.getMessage(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder setResponse(
-          int index, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse value) {
-        if (responseBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureResponseIsMutable();
-          response_.set(index, value);
-          onChanged();
-        } else {
-          responseBuilder_.setMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder setResponse(
-          int index, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder builderForValue) {
-        if (responseBuilder_ == null) {
-          ensureResponseIsMutable();
-          response_.set(index, builderForValue.build());
-          onChanged();
-        } else {
-          responseBuilder_.setMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder addResponse(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse value) {
-        if (responseBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureResponseIsMutable();
-          response_.add(value);
-          onChanged();
-        } else {
-          responseBuilder_.addMessage(value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder addResponse(
-          int index, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse value) {
-        if (responseBuilder_ == null) {
-          if (value == null) {
-            throw new NullPointerException();
-          }
-          ensureResponseIsMutable();
-          response_.add(index, value);
-          onChanged();
-        } else {
-          responseBuilder_.addMessage(index, value);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder addResponse(
-          org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder builderForValue) {
-        if (responseBuilder_ == null) {
-          ensureResponseIsMutable();
-          response_.add(builderForValue.build());
-          onChanged();
-        } else {
-          responseBuilder_.addMessage(builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder addResponse(
-          int index, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder builderForValue) {
-        if (responseBuilder_ == null) {
-          ensureResponseIsMutable();
-          response_.add(index, builderForValue.build());
-          onChanged();
-        } else {
-          responseBuilder_.addMessage(index, builderForValue.build());
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder addAllResponse(
-          java.lang.Iterable<? extends org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> values) {
-        if (responseBuilder_ == null) {
-          ensureResponseIsMutable();
-          super.addAll(values, response_);
-          onChanged();
-        } else {
-          responseBuilder_.addAllMessages(values);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder clearResponse() {
-        if (responseBuilder_ == null) {
-          response_ = java.util.Collections.emptyList();
-          bitField0_ = (bitField0_ & ~0x00000001);
-          onChanged();
-        } else {
-          responseBuilder_.clear();
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public Builder removeResponse(int index) {
-        if (responseBuilder_ == null) {
-          ensureResponseIsMutable();
-          response_.remove(index);
-          onChanged();
-        } else {
-          responseBuilder_.remove(index);
-        }
-        return this;
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder getResponseBuilder(
-          int index) {
-        return getResponseFieldBuilder().getBuilder(index);
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder getResponseOrBuilder(
-          int index) {
-        if (responseBuilder_ == null) {
-          return response_.get(index);  } else {
-          return responseBuilder_.getMessageOrBuilder(index);
-        }
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public java.util.List<? extends org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder> 
-           getResponseOrBuilderList() {
-        if (responseBuilder_ != null) {
-          return responseBuilder_.getMessageOrBuilderList();
-        } else {
-          return java.util.Collections.unmodifiableList(response_);
-        }
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder addResponseBuilder() {
-        return getResponseFieldBuilder().addBuilder(
-            org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder addResponseBuilder(
-          int index) {
-        return getResponseFieldBuilder().addBuilder(
-            index, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.getDefaultInstance());
-      }
-      /**
-       * <code>repeated .commanding.BulkGetCommandResponse.CommandInfoResponse response = 1;</code>
-       */
-      public java.util.List<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder> 
-           getResponseBuilderList() {
-        return getResponseFieldBuilder().getBuilderList();
-      }
-      private com.google.protobuf.RepeatedFieldBuilder<
-          org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder> 
-          getResponseFieldBuilder() {
-        if (responseBuilder_ == null) {
-          responseBuilder_ = new com.google.protobuf.RepeatedFieldBuilder<
-              org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponseOrBuilder>(
-                  response_,
-                  ((bitField0_ & 0x00000001) == 0x00000001),
-                  getParentForChildren(),
-                  isClean());
-          response_ = null;
-        }
-        return responseBuilder_;
-      }
-
-      // @@protoc_insertion_point(builder_scope:commanding.BulkGetCommandResponse)
-    }
-
-    static {
-      defaultInstance = new BulkGetCommandResponse(true);
-      defaultInstance.initFields();
-    }
-
-    // @@protoc_insertion_point(class_scope:commanding.BulkGetCommandResponse)
-  }
-
   private static com.google.protobuf.Descriptors.Descriptor
     internal_static_commanding_CommandId_descriptor;
   private static
@@ -12198,31 +8743,6 @@ public final class Commanding {
   private static
     com.google.protobuf.GeneratedMessage.FieldAccessorTable
       internal_static_commanding_ArgumentAssignmentType_fieldAccessorTable;
-  private static com.google.protobuf.Descriptors.Descriptor
-    internal_static_commanding_ValidateCommandRequest_descriptor;
-  private static
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_commanding_ValidateCommandRequest_fieldAccessorTable;
-  private static com.google.protobuf.Descriptors.Descriptor
-    internal_static_commanding_ValidateCommandResponse_descriptor;
-  private static
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_commanding_ValidateCommandResponse_fieldAccessorTable;
-  private static com.google.protobuf.Descriptors.Descriptor
-    internal_static_commanding_SendCommandRequest_descriptor;
-  private static
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_commanding_SendCommandRequest_fieldAccessorTable;
-  private static com.google.protobuf.Descriptors.Descriptor
-    internal_static_commanding_BulkGetCommandResponse_descriptor;
-  private static
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_commanding_BulkGetCommandResponse_fieldAccessorTable;
-  private static com.google.protobuf.Descriptors.Descriptor
-    internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_descriptor;
-  private static
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_fieldAccessorTable;
 
   public static com.google.protobuf.Descriptors.FileDescriptor
       getDescriptor() {
@@ -12261,24 +8781,14 @@ public final class Commanding {
       " \n\002id\030\003 \001(\0132\024.yamcs.NamedObjectId\0225\n\targ" +
       "uments\030\004 \003(\0132\".commanding.ArgumentAssign" +
       "mentType\"5\n\026ArgumentAssignmentType\022\014\n\004na" +
-      "me\030\001 \001(\t\022\r\n\005value\030\002 \001(\t\"B\n\026ValidateComma",
-      "ndRequest\022(\n\007command\030\001 \003(\0132\027.commanding." +
-      "CommandType\"W\n\027ValidateCommandResponse\022<" +
-      "\n\023commandSignificance\030\001 \003(\0132\037.commanding" +
-      ".CommandSignificance\">\n\022SendCommandReque" +
-      "st\022(\n\007command\030\001 \003(\0132\027.commanding.Command" +
-      "Type\"\276\001\n\026BulkGetCommandResponse\022H\n\010respo" +
-      "nse\030\001 \003(\01326.commanding.BulkGetCommandRes" +
-      "ponse.CommandInfoResponse\032Z\n\023CommandInfo" +
-      "Response\022 \n\002id\030\001 \001(\0132\024.yamcs.NamedObject" +
-      "Id\022!\n\007command\030\002 \001(\0132\020.mdb.CommandInfo*4\n",
-      "\nQueueState\022\013\n\007BLOCKED\020\001\022\014\n\010DISABLED\020\002\022\013" +
-      "\n\007ENABLED\020\0032\313\001\n\023CommandQueueControl\022:\n\rS" +
-      "etQueueState\022\034.commanding.CommandQueueIn" +
-      "fo\032\013.yamcs.Void\022;\n\013SendCommand\022\037.command" +
-      "ing.CommandQueueRequest\032\013.yamcs.Void\022;\n\r" +
-      "RejectCommand\022\035.commanding.CommandQueueE" +
-      "ntry\032\013.yamcs.VoidB\024\n\022org.yamcs.protobuf"
+      "me\030\001 \001(\t\022\r\n\005value\030\002 \001(\t*4\n\nQueueState\022\013\n",
+      "\007BLOCKED\020\001\022\014\n\010DISABLED\020\002\022\013\n\007ENABLED\020\0032\313\001" +
+      "\n\023CommandQueueControl\022:\n\rSetQueueState\022\034" +
+      ".commanding.CommandQueueInfo\032\013.yamcs.Voi" +
+      "d\022;\n\013SendCommand\022\037.commanding.CommandQue" +
+      "ueRequest\032\013.yamcs.Void\022;\n\rRejectCommand\022" +
+      "\035.commanding.CommandQueueEntry\032\013.yamcs.V" +
+      "oidB\024\n\022org.yamcs.protobuf"
     };
     com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
       new com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner() {
@@ -12339,36 +8849,6 @@ public final class Commanding {
             com.google.protobuf.GeneratedMessage.FieldAccessorTable(
               internal_static_commanding_ArgumentAssignmentType_descriptor,
               new java.lang.String[] { "Name", "Value", });
-          internal_static_commanding_ValidateCommandRequest_descriptor =
-            getDescriptor().getMessageTypes().get(9);
-          internal_static_commanding_ValidateCommandRequest_fieldAccessorTable = new
-            com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-              internal_static_commanding_ValidateCommandRequest_descriptor,
-              new java.lang.String[] { "Command", });
-          internal_static_commanding_ValidateCommandResponse_descriptor =
-            getDescriptor().getMessageTypes().get(10);
-          internal_static_commanding_ValidateCommandResponse_fieldAccessorTable = new
-            com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-              internal_static_commanding_ValidateCommandResponse_descriptor,
-              new java.lang.String[] { "CommandSignificance", });
-          internal_static_commanding_SendCommandRequest_descriptor =
-            getDescriptor().getMessageTypes().get(11);
-          internal_static_commanding_SendCommandRequest_fieldAccessorTable = new
-            com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-              internal_static_commanding_SendCommandRequest_descriptor,
-              new java.lang.String[] { "Command", });
-          internal_static_commanding_BulkGetCommandResponse_descriptor =
-            getDescriptor().getMessageTypes().get(12);
-          internal_static_commanding_BulkGetCommandResponse_fieldAccessorTable = new
-            com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-              internal_static_commanding_BulkGetCommandResponse_descriptor,
-              new java.lang.String[] { "Response", });
-          internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_descriptor =
-            internal_static_commanding_BulkGetCommandResponse_descriptor.getNestedTypes().get(0);
-          internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_fieldAccessorTable = new
-            com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-              internal_static_commanding_BulkGetCommandResponse_CommandInfoResponse_descriptor,
-              new java.lang.String[] { "Id", "Command", });
           return null;
         }
       };
