+++ yamcs-api/src/main/java/org/yamcs/protobuf/SchemaCommanding.java
@@ -1207,579 +1207,4 @@ public final class SchemaCommanding
         }
     }
 
-    public static final class ValidateCommandRequest
-    {
-        public static final org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.MessageSchema WRITE =
-            new org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.MessageSchema();
-        public static final org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.BuilderSchema MERGE =
-            new org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.BuilderSchema();
-        
-        public static class MessageSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.ValidateCommandRequest>
-        {
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.ValidateCommandRequest message) throws java.io.IOException
-            {
-                for(org.yamcs.protobuf.Commanding.CommandType command : message.getCommandList())
-                    output.writeObject(1, command, org.yamcs.protobuf.SchemaCommanding.CommandType.WRITE, true);
-
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.ValidateCommandRequest message)
-            {
-                return message.isInitialized();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.ValidateCommandRequest> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.class.getName();
-            }
-            //unused
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.ValidateCommandRequest message) throws java.io.IOException {}
-            public org.yamcs.protobuf.Commanding.ValidateCommandRequest newMessage() { return null; }
-        }
-        public static class BuilderSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder>
-        {
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder builder) throws java.io.IOException
-            {
-                for(int number = input.readFieldNumber(this);; number = input.readFieldNumber(this))
-                {
-                    switch(number)
-                    {
-                        case 0:
-                            return;
-                        case 1:
-                            builder.addCommand(input.mergeObject(org.yamcs.protobuf.Commanding.CommandType.newBuilder(), org.yamcs.protobuf.SchemaCommanding.CommandType.MERGE));
-
-                            break;
-                        default:
-                            input.handleUnknownField(number, this);
-                    }
-                }
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder builder)
-            {
-                return builder.isInitialized();
-            }
-            public org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder newMessage()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.newBuilder();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandRequest.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandRequest.class.getName();
-            }
-            //unused
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.ValidateCommandRequest.Builder builder) throws java.io.IOException {}
-        }
-        public static java.lang.String getFieldName(int number)
-        {
-            switch(number)
-            {
-                case 1: return "command";
-                default: return null;
-            }
-        }
-        public static int getFieldNumber(java.lang.String name)
-        {
-            java.lang.Integer number = fieldMap.get(name);
-            return number == null ? 0 : number.intValue();
-        }
-        private static final java.util.HashMap<java.lang.String,java.lang.Integer> fieldMap = new java.util.HashMap<java.lang.String,java.lang.Integer>();
-        static
-        {
-            fieldMap.put("command", 1);
-        }
-    }
-
-    public static final class ValidateCommandResponse
-    {
-        public static final org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.MessageSchema WRITE =
-            new org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.MessageSchema();
-        public static final org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.BuilderSchema MERGE =
-            new org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.BuilderSchema();
-        
-        public static class MessageSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.ValidateCommandResponse>
-        {
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.ValidateCommandResponse message) throws java.io.IOException
-            {
-                for(org.yamcs.protobuf.Commanding.CommandSignificance commandSignificance : message.getCommandSignificanceList())
-                    output.writeObject(1, commandSignificance, org.yamcs.protobuf.SchemaCommanding.CommandSignificance.WRITE, true);
-
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.ValidateCommandResponse message)
-            {
-                return message.isInitialized();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.ValidateCommandResponse> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.class.getName();
-            }
-            //unused
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.ValidateCommandResponse message) throws java.io.IOException {}
-            public org.yamcs.protobuf.Commanding.ValidateCommandResponse newMessage() { return null; }
-        }
-        public static class BuilderSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder>
-        {
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder builder) throws java.io.IOException
-            {
-                for(int number = input.readFieldNumber(this);; number = input.readFieldNumber(this))
-                {
-                    switch(number)
-                    {
-                        case 0:
-                            return;
-                        case 1:
-                            builder.addCommandSignificance(input.mergeObject(org.yamcs.protobuf.Commanding.CommandSignificance.newBuilder(), org.yamcs.protobuf.SchemaCommanding.CommandSignificance.MERGE));
-
-                            break;
-                        default:
-                            input.handleUnknownField(number, this);
-                    }
-                }
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder builder)
-            {
-                return builder.isInitialized();
-            }
-            public org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder newMessage()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.newBuilder();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.ValidateCommandResponse.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.ValidateCommandResponse.class.getName();
-            }
-            //unused
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.ValidateCommandResponse.Builder builder) throws java.io.IOException {}
-        }
-        public static java.lang.String getFieldName(int number)
-        {
-            switch(number)
-            {
-                case 1: return "commandSignificance";
-                default: return null;
-            }
-        }
-        public static int getFieldNumber(java.lang.String name)
-        {
-            java.lang.Integer number = fieldMap.get(name);
-            return number == null ? 0 : number.intValue();
-        }
-        private static final java.util.HashMap<java.lang.String,java.lang.Integer> fieldMap = new java.util.HashMap<java.lang.String,java.lang.Integer>();
-        static
-        {
-            fieldMap.put("commandSignificance", 1);
-        }
-    }
-
-    public static final class SendCommandRequest
-    {
-        public static final org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.MessageSchema WRITE =
-            new org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.MessageSchema();
-        public static final org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.BuilderSchema MERGE =
-            new org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.BuilderSchema();
-        
-        public static class MessageSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.SendCommandRequest>
-        {
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.SendCommandRequest message) throws java.io.IOException
-            {
-                for(org.yamcs.protobuf.Commanding.CommandType command : message.getCommandList())
-                    output.writeObject(1, command, org.yamcs.protobuf.SchemaCommanding.CommandType.WRITE, true);
-
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.SendCommandRequest message)
-            {
-                return message.isInitialized();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.SendCommandRequest> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.class.getName();
-            }
-            //unused
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.SendCommandRequest message) throws java.io.IOException {}
-            public org.yamcs.protobuf.Commanding.SendCommandRequest newMessage() { return null; }
-        }
-        public static class BuilderSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.SendCommandRequest.Builder>
-        {
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.SendCommandRequest.Builder builder) throws java.io.IOException
-            {
-                for(int number = input.readFieldNumber(this);; number = input.readFieldNumber(this))
-                {
-                    switch(number)
-                    {
-                        case 0:
-                            return;
-                        case 1:
-                            builder.addCommand(input.mergeObject(org.yamcs.protobuf.Commanding.CommandType.newBuilder(), org.yamcs.protobuf.SchemaCommanding.CommandType.MERGE));
-
-                            break;
-                        default:
-                            input.handleUnknownField(number, this);
-                    }
-                }
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.SendCommandRequest.Builder builder)
-            {
-                return builder.isInitialized();
-            }
-            public org.yamcs.protobuf.Commanding.SendCommandRequest.Builder newMessage()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.newBuilder();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.SendCommandRequest.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.SendCommandRequest.Builder> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.Builder.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.SendCommandRequest.class.getName();
-            }
-            //unused
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.SendCommandRequest.Builder builder) throws java.io.IOException {}
-        }
-        public static java.lang.String getFieldName(int number)
-        {
-            switch(number)
-            {
-                case 1: return "command";
-                default: return null;
-            }
-        }
-        public static int getFieldNumber(java.lang.String name)
-        {
-            java.lang.Integer number = fieldMap.get(name);
-            return number == null ? 0 : number.intValue();
-        }
-        private static final java.util.HashMap<java.lang.String,java.lang.Integer> fieldMap = new java.util.HashMap<java.lang.String,java.lang.Integer>();
-        static
-        {
-            fieldMap.put("command", 1);
-        }
-    }
-
-    public static final class BulkGetCommandResponse
-    {
-
-        public static final class CommandInfoResponse
-        {
-            public static final org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.MessageSchema WRITE =
-                new org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.MessageSchema();
-            public static final org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.BuilderSchema MERGE =
-                new org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.BuilderSchema();
-            
-            public static class MessageSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse>
-            {
-                public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse message) throws java.io.IOException
-                {
-                    if(message.hasId())
-                        output.writeObject(1, message.getId(), org.yamcs.protobuf.SchemaYamcs.NamedObjectId.WRITE, false);
-
-                    if(message.hasCommand())
-                        output.writeObject(2, message.getCommand(), org.yamcs.protobuf.SchemaMdb.CommandInfo.WRITE, false);
-
-                }
-                public boolean isInitialized(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse message)
-                {
-                    return message.isInitialized();
-                }
-                public java.lang.String getFieldName(int number)
-                {
-                    return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.getFieldName(number);
-                }
-                public int getFieldNumber(java.lang.String name)
-                {
-                    return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.getFieldNumber(name);
-                }
-                public java.lang.Class<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse> typeClass()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class;
-                }
-                public java.lang.String messageName()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class.getSimpleName();
-                }
-                public java.lang.String messageFullName()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class.getName();
-                }
-                //unused
-                public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse message) throws java.io.IOException {}
-                public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse newMessage() { return null; }
-            }
-            public static class BuilderSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder>
-            {
-                public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder builder) throws java.io.IOException
-                {
-                    for(int number = input.readFieldNumber(this);; number = input.readFieldNumber(this))
-                    {
-                        switch(number)
-                        {
-                            case 0:
-                                return;
-                            case 1:
-                                builder.setId(input.mergeObject(org.yamcs.protobuf.Yamcs.NamedObjectId.newBuilder(), org.yamcs.protobuf.SchemaYamcs.NamedObjectId.MERGE));
-
-                                break;
-                            case 2:
-                                builder.setCommand(input.mergeObject(org.yamcs.protobuf.Mdb.CommandInfo.newBuilder(), org.yamcs.protobuf.SchemaMdb.CommandInfo.MERGE));
-
-                                break;
-                            default:
-                                input.handleUnknownField(number, this);
-                        }
-                    }
-                }
-                public boolean isInitialized(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder builder)
-                {
-                    return builder.isInitialized();
-                }
-                public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder newMessage()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.newBuilder();
-                }
-                public java.lang.String getFieldName(int number)
-                {
-                    return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.getFieldName(number);
-                }
-                public int getFieldNumber(java.lang.String name)
-                {
-                    return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.getFieldNumber(name);
-                }
-                public java.lang.Class<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder> typeClass()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder.class;
-                }
-                public java.lang.String messageName()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class.getSimpleName();
-                }
-                public java.lang.String messageFullName()
-                {
-                    return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.class.getName();
-                }
-                //unused
-                public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.Builder builder) throws java.io.IOException {}
-            }
-            public static java.lang.String getFieldName(int number)
-            {
-                switch(number)
-                {
-                    case 1: return "id";
-                    case 2: return "command";
-                    default: return null;
-                }
-            }
-            public static int getFieldNumber(java.lang.String name)
-            {
-                java.lang.Integer number = fieldMap.get(name);
-                return number == null ? 0 : number.intValue();
-            }
-            private static final java.util.HashMap<java.lang.String,java.lang.Integer> fieldMap = new java.util.HashMap<java.lang.String,java.lang.Integer>();
-            static
-            {
-                fieldMap.put("id", 1);
-                fieldMap.put("command", 2);
-            }
-        }
-
-        public static final org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.MessageSchema WRITE =
-            new org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.MessageSchema();
-        public static final org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.BuilderSchema MERGE =
-            new org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.BuilderSchema();
-        
-        public static class MessageSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.BulkGetCommandResponse>
-        {
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.BulkGetCommandResponse message) throws java.io.IOException
-            {
-                for(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse response : message.getResponseList())
-                    output.writeObject(1, response, org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.WRITE, true);
-
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.BulkGetCommandResponse message)
-            {
-                return message.isInitialized();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.BulkGetCommandResponse> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class.getName();
-            }
-            //unused
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.BulkGetCommandResponse message) throws java.io.IOException {}
-            public org.yamcs.protobuf.Commanding.BulkGetCommandResponse newMessage() { return null; }
-        }
-        public static class BuilderSchema implements io.protostuff.Schema<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder>
-        {
-            public void mergeFrom(io.protostuff.Input input, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder builder) throws java.io.IOException
-            {
-                for(int number = input.readFieldNumber(this);; number = input.readFieldNumber(this))
-                {
-                    switch(number)
-                    {
-                        case 0:
-                            return;
-                        case 1:
-                            builder.addResponse(input.mergeObject(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.CommandInfoResponse.newBuilder(), org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.CommandInfoResponse.MERGE));
-
-                            break;
-                        default:
-                            input.handleUnknownField(number, this);
-                    }
-                }
-            }
-            public boolean isInitialized(org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder builder)
-            {
-                return builder.isInitialized();
-            }
-            public org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder newMessage()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.newBuilder();
-            }
-            public java.lang.String getFieldName(int number)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.getFieldName(number);
-            }
-            public int getFieldNumber(java.lang.String name)
-            {
-                return org.yamcs.protobuf.SchemaCommanding.BulkGetCommandResponse.getFieldNumber(name);
-            }
-            public java.lang.Class<org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder> typeClass()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder.class;
-            }
-            public java.lang.String messageName()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class.getSimpleName();
-            }
-            public java.lang.String messageFullName()
-            {
-                return org.yamcs.protobuf.Commanding.BulkGetCommandResponse.class.getName();
-            }
-            //unused
-            public void writeTo(io.protostuff.Output output, org.yamcs.protobuf.Commanding.BulkGetCommandResponse.Builder builder) throws java.io.IOException {}
-        }
-        public static java.lang.String getFieldName(int number)
-        {
-            switch(number)
-            {
-                case 1: return "response";
-                default: return null;
-            }
-        }
-        public static int getFieldNumber(java.lang.String name)
-        {
-            java.lang.Integer number = fieldMap.get(name);
-            return number == null ? 0 : number.intValue();
-        }
-        private static final java.util.HashMap<java.lang.String,java.lang.Integer> fieldMap = new java.util.HashMap<java.lang.String,java.lang.Integer>();
-        static
-        {
-            fieldMap.put("response", 1);
-        }
-    }
-
 }
