+++ yamcs-core/src/test/java/org/yamcs/AbstractIntegrationTest.java
@@ -27,8 +27,8 @@ import org.yamcs.protobuf.Alarms.Alarm;
 import org.yamcs.protobuf.Commanding.ArgumentAssignmentType;
 import org.yamcs.protobuf.Commanding.CommandHistoryEntry;
 import org.yamcs.protobuf.Commanding.CommandType;
-import org.yamcs.protobuf.Commanding.SendCommandRequest;
 import org.yamcs.protobuf.Pvalue.ParameterData;
+import org.yamcs.protobuf.Rest.SendCommandRequest;
 import org.yamcs.protobuf.Web.WebSocketServerMessage.WebSocketSubscriptionData;
 import org.yamcs.protobuf.Yamcs.Event;
 import org.yamcs.protobuf.Yamcs.NamedObjectId;
