+++ yamcs-core/src/test/java/org/yamcs/PermissionsTest.java
@@ -10,12 +10,11 @@ import org.junit.Test;
 import org.yamcs.api.ws.WebSocketRequest;
 import org.yamcs.protobuf.Archive.DumpArchiveRequest;
 import org.yamcs.protobuf.Archive.DumpArchiveResponse;
-import org.yamcs.protobuf.Commanding.SendCommandRequest;
 import org.yamcs.protobuf.Pvalue.ParameterData;
 import org.yamcs.protobuf.Pvalue.ParameterValue;
 import org.yamcs.protobuf.Rest.BulkGetParameterValueRequest;
+import org.yamcs.protobuf.Rest.SendCommandRequest;
 import org.yamcs.protobuf.SchemaArchive;
-import org.yamcs.protobuf.SchemaCommanding;
 import org.yamcs.protobuf.SchemaPvalue;
 import org.yamcs.protobuf.SchemaRest;
 import org.yamcs.protobuf.ValueHelper;
@@ -39,7 +38,7 @@ public class PermissionsTest extends AbstractIntegrationTest {
         ParameterReplayRequest prr = ParameterReplayRequest.newBuilder().addNameFilter(p1_1_6id).addNameFilter(p1_3_1id).build();
         DumpArchiveRequest dumpRequest = DumpArchiveRequest.newBuilder().setParameterRequest(prr)
                 .setUtcStart("2016-02-03T10:10:00").setUtcStop("2016-02-03T10:10:02").build();
-        String response = httpClient.doGetRequest("http://localhost:9190/IntegrationTest/api/archive", toJson(dumpRequest, SchemaArchive.DumpArchiveRequest.WRITE), currentUser);
+        String response = httpClient.doGetRequest("http://localhost:9190/api/IntegrationTest/archive", toJson(dumpRequest, SchemaArchive.DumpArchiveRequest.WRITE), currentUser);
         DumpArchiveResponse rdar = (fromJson(response, SchemaArchive.DumpArchiveResponse.MERGE)).build();
         List<ParameterData> plist = rdar.getParameterDataList();
         assertNotNull(plist);
@@ -79,7 +78,7 @@ public class PermissionsTest extends AbstractIntegrationTest {
         ParameterReplayRequest prr = ParameterReplayRequest.newBuilder().addNameFilter(p1_1_6id).build();
         DumpArchiveRequest dumpRequest = DumpArchiveRequest.newBuilder().setParameterRequest(prr)
                 .setUtcStart("2015-03-02T10:10:00").setUtcStop("2015-03-02T10:10:02").build();
-        String response = httpClient.doGetRequest("http://localhost:9190/IntegrationTest/api/archive", toJson(dumpRequest, SchemaArchive.DumpArchiveRequest.WRITE), currentUser);
+        String response = httpClient.doGetRequest("http://localhost:9190/api/IntegrationTest/archive", toJson(dumpRequest, SchemaArchive.DumpArchiveRequest.WRITE), currentUser);
         DumpArchiveResponse rdar = (fromJson(response, SchemaArchive.DumpArchiveResponse.MERGE)).build();
         List<ParameterData> plist = rdar.getParameterDataList();
         assertNotNull(plist);
@@ -95,7 +94,7 @@ public class PermissionsTest extends AbstractIntegrationTest {
             prr = ParameterReplayRequest.newBuilder().addNameFilter(p1_3_1id).build();
             dumpRequest = DumpArchiveRequest.newBuilder().setParameterRequest(prr)
                     .setUtcStart("2015-03-02T10:10:00").setUtcStop("2015-03-02T10:10:02").build();
-            response = httpClient.doGetRequest("http://localhost:9190/IntegrationTest/api/archive", toJson(dumpRequest, SchemaArchive.DumpArchiveRequest.WRITE), currentUser);
+            response = httpClient.doGetRequest("http://localhost:9190/api/IntegrationTest/archive", toJson(dumpRequest, SchemaArchive.DumpArchiveRequest.WRITE), currentUser);
             rdar = (fromJson(response, SchemaArchive.DumpArchiveResponse.MERGE)).build();
             plist = rdar.getParameterDataList();
             if(plist.size() == 0) {
@@ -118,12 +117,12 @@ public class PermissionsTest extends AbstractIntegrationTest {
         WebSocketRequest wsr = new WebSocketRequest("cmdhistory", "subscribe");
         wsClient.sendRequest(wsr);
         SendCommandRequest cmdreq = getCommand("/REFMDB/SUBSYS1/INT_ARG_TC", 5, "uint32_arg", "1000");
-        String resp = httpClient.doRequest("http://localhost:9190/IntegrationTest/api/commanding/queue", HttpMethod.POST, toJson(cmdreq, SchemaCommanding.SendCommandRequest.WRITE), currentUser);
+        String resp = httpClient.doRequest("http://localhost:9190/api/IntegrationTest/commanding/queue", HttpMethod.POST, toJson(cmdreq, SchemaRest.SendCommandRequest.WRITE), currentUser);
         assertEquals("", resp);
 
         // Command FLOAT_ARG_TC is denied
         cmdreq = getCommand("/REFMDB/SUBSYS1/FLOAT_ARG_TC", 5, "float_arg", "-15", "double_arg", "0");
-        resp = httpClient.doRequest("http://localhost:9190/IntegrationTest/api/commanding/queue", HttpMethod.POST, toJson(cmdreq, SchemaCommanding.SendCommandRequest.WRITE), currentUser);
+        resp = httpClient.doRequest("http://localhost:9190/api/IntegrationTest/commanding/queue", HttpMethod.POST, toJson(cmdreq, SchemaRest.SendCommandRequest.WRITE), currentUser);
         assertTrue("Should get permission exception message", resp.contains("ForbiddenException"));
     }
 
@@ -135,13 +134,13 @@ public class PermissionsTest extends AbstractIntegrationTest {
         // Allowed to subscribe to Integer parameter from cache
         NamedObjectList validSubscrList = getSubscription("/REFMDB/SUBSYS1/IntegerPara1_1_6", "/REFMDB/SUBSYS1/IntegerPara1_1_7");
         BulkGetParameterValueRequest req = BulkGetParameterValueRequest.newBuilder().setFromCache(true).addAllId(validSubscrList.getListList()).build();
-        String response = httpClient.doRequest("http://localhost:9190/IntegrationTest/api/parameter/_get", HttpMethod.GET, toJson(req, SchemaRest.BulkGetParameterValueRequest.WRITE), currentUser);
+        String response = httpClient.doRequest("http://localhost:9190/api/IntegrationTest/parameter/_get", HttpMethod.GET, toJson(req, SchemaRest.BulkGetParameterValueRequest.WRITE), currentUser);
         assertTrue("{}", !response.contains("ForbiddenException"));
 
         // Denied to subscribe to Float parameter from cache
         validSubscrList = getSubscription("/REFMDB/SUBSYS1/FloatPara1_1_3", "/REFMDB/SUBSYS1/FloatPara1_1_2");
         req = BulkGetParameterValueRequest.newBuilder().setFromCache(true).addAllId(validSubscrList.getListList()).build();
-        response = httpClient.doRequest("http://localhost:9190/IntegrationTest/api/parameter/_get", HttpMethod.GET, toJson(req, SchemaRest.BulkGetParameterValueRequest.WRITE), currentUser);
+        response = httpClient.doRequest("http://localhost:9190/api/IntegrationTest/parameter/_get", HttpMethod.GET, toJson(req, SchemaRest.BulkGetParameterValueRequest.WRITE), currentUser);
         assertTrue("Permission should be denied", response.contains("ForbiddenException"));
     }
 
@@ -155,7 +154,7 @@ public class PermissionsTest extends AbstractIntegrationTest {
                 .setEngValue(ValueHelper.newValue(5)).build();
         ParameterData pdata = ParameterData.newBuilder().addParameter(pv1).build();
         HttpClient httpClient = new HttpClient();
-        String response = httpClient.doRequest("http://localhost:9190/IntegrationTest/api/parameter/_set", HttpMethod.POST, toJson(pdata, SchemaPvalue.ParameterData.WRITE), currentUser);
+        String response = httpClient.doRequest("http://localhost:9190/api/IntegrationTest/parameter/_set", HttpMethod.POST, toJson(pdata, SchemaPvalue.ParameterData.WRITE), currentUser);
         assertTrue("Permission should be denied", response.contains("ForbiddenException"));
     }
